refactor(efi): split boot_services.rs into a module directory
Summary:
  - Move the 2714-line table module into boot_services/{mod,events,
    memory,protocols_db,images}.rs with zero logic changes: every line
    lands verbatim except visibility (pub(super) on table-referenced
    fns), qualified table field paths, one-level super:: adjustments,
    and per-file import headers
  - Event-type consts go private in events.rs (one cfg(test)-gated,
    one pub(super) for the legacy ExitBootServices block); the two ID
    consts are re-exported so boot_services::KEYBOARD_EVENT_ID and
    POINTER_EVENT_ID paths are unchanged, as are create_handle,
    install_protocol, measure_*, and serialize_* paths
  - Drop the stale mid-file guid_fmt import (now a top import in
    protocols_db) and its outdated extraction comment

Tests:
  - cargo fmt --check
  - cargo clippy --workspace --release -- -D warnings; ui and fb-log
    features; aarch64 and riscv64 targets
  - cargo test --workspace (event unit tests moved with the code)
  - ./crabefi build for x86-64, aarch64 (virt/sbsa), riscv64
  - xtask tests, coreboot feature script
