--- crabefi-core/src/efi/boot_services.rs
@@ -1,3251 +1,0 @@
-//! EFI Boot Services
-//!
-//! This module implements the EFI Boot Services table, which provides
-//! memory allocation, protocol handling, and image loading services.
-//!
-//! # State Management
-//!
-//! Boot Services state (handles, events, loaded images) lives in
-//! [`super::tables`]. Access it via `tables()` and `with_tables_mut()`.
-
-#[cfg(test)]
-#[path = "boot_services_tests.rs"]
-mod protocol_tests;
-
-use super::allocator::{self, AllocateType, MemoryDescriptor, MemoryType};
-use super::image_loader;
-use super::protocols::loaded_image::{LOADED_IMAGE_PROTOCOL_GUID, create_loaded_image_protocol};
-use super::system_table;
-use super::tables::{
-    EventEntry, HandleEntry, LoadedImageEntry, MAX_EVENTS, MAX_PROTOCOL_NOTIFIES,
-    MAX_PROTOCOLS_PER_HANDLE, OpenProtocolEntry, ProtocolEntry, ProtocolNotifyEntry, Tables,
-    TimerType, tables, with_tables_mut,
-};
-use crate::cell::StaticMut;
-use crate::pe;
-use alloc::vec::Vec;
-use core::ffi::c_void;
-
-use crabefi_efi_types::crc32;
-use r_efi::efi::{self, Boolean, Guid, Handle, Status, SystemTable, TableHeader, Tpl};
-use r_efi::protocols::device_path::Protocol as DevicePathProtocol;
-
-/// Boot Services signature "BOOTSERV"
-const EFI_BOOT_SERVICES_SIGNATURE: u64 = 0x56524553544F4F42;
-
-/// Boot Services revision (matches system table)
-const EFI_BOOT_SERVICES_REVISION: u32 = (2 << 16) | 100;
-
-/// Event types
-pub const EVT_TIMER: u32 = 0x80000000;
-pub const EVT_RUNTIME: u32 = 0x40000000;
-pub const EVT_NOTIFY_WAIT: u32 = 0x00000100;
-pub const EVT_NOTIFY_SIGNAL: u32 = 0x00000200;
-pub const EVT_SIGNAL_EXIT_BOOT_SERVICES: u32 = 0x00000201;
-pub const EVT_SIGNAL_VIRTUAL_ADDRESS_CHANGE: u32 = 0x60000202;
-
-/// Special event ID for keyboard input
-pub const KEYBOARD_EVENT_ID: usize = 1;
-
-/// Special event ID for pointer (mouse) input
-#[cfg(feature = "ui")]
-pub const POINTER_EVENT_ID: usize = 2;
-
-#[cfg(feature = "ui")]
-const FIRST_DYNAMIC_EVENT_ID: usize = POINTER_EVENT_ID + 1;
-#[cfg(not(feature = "ui"))]
-const FIRST_DYNAMIC_EVENT_ID: usize = KEYBOARD_EVENT_ID + 1;
-
-/// Static boot services table
-static BOOT_SERVICES: StaticMut<efi::BootServices> = StaticMut::new(efi::BootServices {
-    hdr: TableHeader {
-        signature: EFI_BOOT_SERVICES_SIGNATURE,
-        revision: EFI_BOOT_SERVICES_REVISION,
-        header_size: core::mem::size_of::<efi::BootServices>() as u32,
-        crc32: 0,
-        reserved: 0,
-    },
-    raise_tpl,
-    restore_tpl,
-    allocate_pages,
-    free_pages,
-    get_memory_map,
-    allocate_pool,
-    free_pool,
-    create_event,
-    set_timer,
-    wait_for_event,
-    signal_event,
-    close_event,
-    check_event,
-    install_protocol_interface,
-    reinstall_protocol_interface,
-    uninstall_protocol_interface,
-    handle_protocol,
-    reserved: core::ptr::null_mut(),
-    register_protocol_notify,
-    locate_handle,
-    locate_device_path,
-    install_configuration_table,
-    load_image,
-    start_image,
-    exit,
-    unload_image,
-    exit_boot_services,
-    get_next_monotonic_count,
-    stall,
-    set_watchdog_timer,
-    connect_controller,
-    disconnect_controller,
-    open_protocol,
-    close_protocol,
-    open_protocol_information,
-    protocols_per_handle,
-    locate_handle_buffer,
-    locate_protocol,
-    // r-efi's function-pointer aliases omit the variadic tail. These entry
-    // points preserve the real UEFI ABI and parse protocol/interface pairs
-    // through the terminating NULL protocol argument.
-    install_multiple_protocol_interfaces: INSTALL_MULTIPLE_PROTOCOL_INTERFACES,
-    uninstall_multiple_protocol_interfaces: UNINSTALL_MULTIPLE_PROTOCOL_INTERFACES,
-    calculate_crc32,
-    copy_mem,
-    set_mem,
-    create_event_ex,
-});
-
-/// Get a pointer to the boot services table
-pub fn get_boot_services() -> *mut efi::BootServices {
-    BOOT_SERVICES.get()
-}
-
-// ============================================================================
-// TPL (Task Priority Level) Functions
-// ============================================================================
-
-extern "efiapi" fn raise_tpl(new_tpl: Tpl) -> Tpl {
-    log::debug!("BS.RaiseTpl({:?})", new_tpl);
-    // No interrupt handling, return current TPL (APPLICATION)
-    efi::TPL_APPLICATION
-}
-
-extern "efiapi" fn restore_tpl(old_tpl: Tpl) {
-    log::debug!("BS.RestoreTpl({:?})", old_tpl);
-    // No-op
-}
-
-// ============================================================================
-// Memory Allocation Functions
-// ============================================================================
-
-extern "efiapi" fn allocate_pages(
-    alloc_type: efi::AllocateType,
-    memory_type: efi::MemoryType,
-    pages: usize,
-    memory: *mut efi::PhysicalAddress,
-) -> Status {
-    log::debug!(
-        "BS.AllocatePages(type={}, mem_type={}, pages={}, addr={:#x})",
-        alloc_type,
-        memory_type,
-        pages,
-        if memory.is_null() {
-            0
-        } else {
-            unsafe { *memory }
-        }
-    );
-
-    if memory.is_null() {
-        return Status::INVALID_PARAMETER;
-    }
-
-    let alloc_type = match AllocateType::try_from(alloc_type) {
-        Ok(t) => t,
-        Err(e) => {
-            log::debug!("BS.AllocatePages: {e}");
-            return Status::INVALID_PARAMETER;
-        }
-    };
-
-    let mem_type = match MemoryType::try_from(memory_type) {
-        Ok(t) => t,
-        Err(e) => {
-            log::debug!("BS.AllocatePages: {e}");
-            return Status::INVALID_PARAMETER;
-        }
-    };
-
-    let mut addr = unsafe { *memory };
-    let status = allocator::allocate_pages(alloc_type, mem_type, pages as u64, &mut addr);
-
-    if status == Status::SUCCESS {
-        unsafe { *memory = addr };
-        log::debug!("  -> allocated at {:#x}", addr);
-    } else {
-        log::warn!("  -> failed: {:?}", status);
-    }
-
-    status
-}
-
-extern "efiapi" fn free_pages(memory: efi::PhysicalAddress, pages: usize) -> Status {
-    allocator::free_pages(memory, pages as u64)
-}
-
-extern "efiapi" fn get_memory_map(
-    memory_map_size: *mut usize,
-    memory_map: *mut efi::MemoryDescriptor,
-    map_key: *mut usize,
-    descriptor_size: *mut usize,
-    descriptor_version: *mut u32,
-) -> Status {
-    log::debug!(
-        "BS.GetMemoryMap(buf_size={:?}, map={:?})",
-        if memory_map_size.is_null() {
-            0
-        } else {
-            unsafe { *memory_map_size }
-        },
-        memory_map
-    );
-
-    if memory_map_size.is_null()
-        || map_key.is_null()
-        || descriptor_size.is_null()
-        || descriptor_version.is_null()
-    {
-        return Status::INVALID_PARAMETER;
-    }
-
-    let mut size = unsafe { *memory_map_size };
-    let mut key = 0usize;
-    let mut desc_size = 0usize;
-    let mut desc_version = 0u32;
-
-    // Convert memory_map pointer to a slice if not null
-    let map_opt = if memory_map.is_null() {
-        None
-    } else {
-        let num_entries = size / core::mem::size_of::<MemoryDescriptor>();
-        Some(unsafe {
-            core::slice::from_raw_parts_mut(memory_map as *mut MemoryDescriptor, num_entries)
-        })
-    };
-
-    let status = allocator::get_memory_map(
-        &mut size,
-        map_opt,
-        &mut key,
-        &mut desc_size,
-        &mut desc_version,
-    );
-
-    unsafe {
-        *memory_map_size = size;
-        *map_key = key;
-        *descriptor_size = desc_size;
-        *descriptor_version = desc_version;
-    }
-
-    log::debug!("  -> {:?} (size={}, key={:#x})", status, size, key);
-    status
-}
-
-extern "efiapi" fn allocate_pool(
-    pool_type: efi::MemoryType,
-    size: usize,
-    buffer: *mut *mut c_void,
-) -> Status {
-    log::trace!("BS.AllocatePool(type={}, size={})", pool_type, size);
-
-    if buffer.is_null() || size == 0 {
-        return Status::INVALID_PARAMETER;
-    }
-
-    let mem_type = match MemoryType::try_from(pool_type) {
-        Ok(t) => t,
-        Err(e) => {
-            log::debug!("BS.AllocatePool: {e}");
-            return Status::INVALID_PARAMETER;
-        }
-    };
-
-    match allocator::allocate_pool(mem_type, size) {
-        Ok(ptr) => {
-            unsafe { *buffer = ptr as *mut c_void };
-            Status::SUCCESS
-        }
-        Err(status) => status,
-    }
-}
-
-extern "efiapi" fn free_pool(buffer: *mut c_void) -> Status {
-    log::trace!("BS.FreePool({:?})", buffer);
-    if buffer.is_null() {
-        return Status::INVALID_PARAMETER;
-    }
-
-    allocator::free_pool(buffer as *mut u8)
-}
-
-// ============================================================================
-// Event Functions (mostly unsupported)
-// ============================================================================
-
-const fn boot_event_type_supported(event_type: u32) -> bool {
-    event_type & EVT_RUNTIME == 0
-}
-
-fn find_free_event_slot(events: &[EventEntry]) -> Option<usize> {
-    events
-        .iter()
-        .enumerate()
-        .skip(FIRST_DYNAMIC_EVENT_ID)
-        .find_map(|(event_id, entry)| (!entry.in_use).then_some(event_id))
-}
-
-fn next_event_generation(generation: usize) -> usize {
-    let max_generation = usize::MAX / MAX_EVENTS;
-    if generation == 0 || generation == max_generation {
-        1
-    } else {
-        generation + 1
-    }
-}
-
-fn event_handle(event_id: usize, generation: usize) -> efi::Event {
-    (generation * MAX_EVENTS + event_id) as *mut c_void
-}
-
-fn event_id_for_handle(events: &[EventEntry], event: efi::Event) -> Option<usize> {
-    let raw = event as usize;
-    if raw == KEYBOARD_EVENT_ID {
-        return Some(KEYBOARD_EVENT_ID);
-    }
-    #[cfg(feature = "ui")]
-    if raw == POINTER_EVENT_ID {
-        return Some(POINTER_EVENT_ID);
-    }
-
-    let event_id = raw % MAX_EVENTS;
-    let generation = raw / MAX_EVENTS;
-    (event_id >= FIRST_DYNAMIC_EVENT_ID
-        && generation != 0
-        && events[event_id].in_use
-        && events[event_id].generation == generation)
-        .then_some(event_id)
-}
-
-fn dynamic_event_id_for_handle(events: &[EventEntry], event: efi::Event) -> Option<usize> {
-    event_id_for_handle(events, event).filter(|event_id| *event_id >= FIRST_DYNAMIC_EVENT_ID)
-}
-
-extern "efiapi" fn create_event(
-    event_type: u32,
-    notify_tpl: Tpl,
-    notify_function: Option<efi::EventNotify>,
-    notify_context: *mut c_void,
-    event: *mut efi::Event,
-) -> Status {
-    log::debug!(
-        "BS.CreateEvent(type={:#x}, tpl={:?})",
-        event_type,
-        notify_tpl
-    );
-
-    if event.is_null() {
-        return Status::INVALID_PARAMETER;
-    }
-    if !boot_event_type_supported(event_type) {
-        return Status::INVALID_PARAMETER;
-    }
-
-    // Allocate a reusable event slot from centralized state.
-    with_tables_mut(|efi_state| {
-        let Some(event_id) = find_free_event_slot(&efi_state.events) else {
-            log::error!("  -> OUT_OF_RESOURCES (no more event slots)");
-            return Status::OUT_OF_RESOURCES;
-        };
-
-        let generation = next_event_generation(efi_state.events[event_id].generation);
-
-        // Store event info including notify callback.
-        efi_state.events[event_id] = EventEntry {
-            in_use: true,
-            generation,
-            event_type,
-            notify_tpl,
-            signaled: false,
-            is_keyboard_event: false,
-            notify_function,
-            notify_context,
-            event_group: None,
-            timer_type: TimerType::Cancel,
-            timer_trigger_time: 0,
-            timer_deadline_tsc: 0,
-        };
-
-        let handle = event_handle(event_id, generation);
-        unsafe {
-            *event = handle;
-        }
-
-        log::debug!(
-            "  -> SUCCESS (event={:?}, slot={}, generation={})",
-            handle,
-            event_id,
-            generation
-        );
-        Status::SUCCESS
-    })
-}
-
-extern "efiapi" fn set_timer(
-    event: efi::Event,
-    timer_type: efi::TimerDelay,
-    trigger_time: u64,
-) -> Status {
-    log::debug!(
-        "BS.SetTimer(event={:?}, type={}, time={})",
-        event,
-        timer_type,
-        trigger_time
-    );
-
-    let timer = match TimerType::try_from(timer_type) {
-        Ok(t) => t,
-        Err(e) => {
-            log::debug!("BS.SetTimer: {e}");
-            return Status::INVALID_PARAMETER;
-        }
-    };
-
-    with_tables_mut(|efi_state| {
-        let Some(event_id) = dynamic_event_id_for_handle(&efi_state.events, event) else {
-            return Status::INVALID_PARAMETER;
-        };
-        let entry = &mut efi_state.events[event_id];
-
-        // Verify this is a timer event
-        if entry.event_type & EVT_TIMER == 0 {
-            log::debug!("  -> INVALID_PARAMETER (not a timer event)");
-            return Status::INVALID_PARAMETER;
-        }
-
-        entry.timer_type = timer;
-        entry.timer_trigger_time = trigger_time;
-
-        match timer {
-            TimerType::Cancel => {
-                entry.timer_deadline_tsc = 0;
-                entry.signaled = false;
-                log::debug!("  -> SUCCESS (timer cancelled)");
-            }
-            TimerType::Periodic | TimerType::Relative => {
-                // Convert 100ns units to TSC ticks
-                let tsc_freq = crate::time::tsc_frequency();
-                if tsc_freq == 0 {
-                    log::error!("  -> DEVICE_ERROR (TSC not calibrated)");
-                    return Status::DEVICE_ERROR;
-                }
-                let tsc_per_us = tsc_freq / 1_000_000;
-                let us = trigger_time / 10;
-                let tsc_offset = us * tsc_per_us.max(1);
-                let now = crate::time::rdtsc();
-                entry.timer_deadline_tsc = now + tsc_offset;
-                log::debug!("  -> SUCCESS (deadline in {}us)", us);
-            }
-        }
-
-        Status::SUCCESS
-    })
-}
-
-fn notify_wait_event(event_id: usize, event: efi::Event) {
-    let notify_fn = with_tables_mut(|efi_state| {
-        let entry = &efi_state.events[event_id];
-        if !entry.signaled && entry.event_type & EVT_NOTIFY_WAIT != 0 {
-            entry.notify_function.map(|f| (f, entry.notify_context))
-        } else {
-            None
-        }
-    });
-
-    if let Some((func, context)) = notify_fn {
-        log::trace!(
-            "  -> Calling EVT_NOTIFY_WAIT function for event {}",
-            event_id
-        );
-        func(event, context);
-    }
-}
-
-extern "efiapi" fn wait_for_event(
-    number_of_events: usize,
-    event: *mut efi::Event,
-    index: *mut usize,
-) -> Status {
-    log::debug!("BS.WaitForEvent(count={})", number_of_events);
-
-    if number_of_events == 0 || event.is_null() || index.is_null() {
-        return Status::INVALID_PARAMETER;
-    }
-
-    // Get the list of events to wait on
-    let events_to_wait = unsafe { core::slice::from_raw_parts(event, number_of_events) };
-
-    // Poll for events (keyboard input, timers, signaled events)
-    loop {
-        // Check each event
-        for (i, &evt) in events_to_wait.iter().enumerate() {
-            let Some(event_id) =
-                with_tables_mut(|efi_state| event_id_for_handle(&efi_state.events, evt))
-            else {
-                return Status::INVALID_PARAMETER;
-            };
-
-            // Check if it's the keyboard event and there's actual key input.
-            // We do a real read-ahead (not just peek at status registers) to
-            // avoid false positives from modifier keys, mouse data, etc.
-            if event_id == KEYBOARD_EVENT_ID
-                && crate::efi::protocols::console::keyboard_check_ready()
-            {
-                unsafe { *index = i };
-                log::debug!("  -> SUCCESS (keyboard input ready, index={})", i);
-                return Status::SUCCESS;
-            }
-
-            // Check if it's the pointer event and there's mouse input.
-            #[cfg(feature = "ui")]
-            if event_id == POINTER_EVENT_ID
-                && crate::efi::protocols::simple_pointer::pointer_check_ready()
-            {
-                unsafe { *index = i };
-                log::debug!("  -> SUCCESS (pointer input ready, index={})", i);
-                return Status::SUCCESS;
-            }
-
-            // Check if a regular dynamic event is signaled, including timers.
-            if event_id >= FIRST_DYNAMIC_EVENT_ID {
-                notify_wait_event(event_id, evt);
-                check_timer_event(event_id);
-
-                // Per UEFI spec: WaitForEvent clears the signaled state
-                // of the event that triggered the return.
-                let signaled = with_tables_mut(|efi_state| {
-                    let was_signaled = efi_state.events[event_id].signaled;
-                    if was_signaled {
-                        efi_state.events[event_id].signaled = false;
-                    }
-                    was_signaled
-                });
-                if signaled {
-                    unsafe { *index = i };
-                    log::debug!("  -> SUCCESS (event signaled, index={})", i);
-                    return Status::SUCCESS;
-                }
-            }
-        }
-
-        // Small delay to avoid busy-waiting too aggressively
-        for _ in 0..1000 {
-            core::hint::spin_loop();
-        }
-    }
-}
-
-/// Pending notify callback extracted when signaling an event.
-type SignaledNotify = Option<(efi::EventNotify, *mut c_void)>;
-
-/// Mark an event signaled and extract its pending `EVT_NOTIFY_SIGNAL` callback.
-///
-/// Factored out of `signal_event` so the lookup/state mutation is unit
-/// testable without global tables. Uses the general handle lookup on purpose:
-/// the handle's static/dynamic classification is an internal detail and must
-/// not narrow public `SignalEvent` semantics — any valid event (including the
-/// static keyboard/pointer events) signals successfully. Only the
-/// notification behavior depends on the event type.
-fn signal_event_entry(
-    events: &mut [EventEntry],
-    event: efi::Event,
-) -> Result<(usize, SignaledNotify), Status> {
-    let Some(event_id) = event_id_for_handle(events, event) else {
-        return Err(Status::INVALID_PARAMETER);
-    };
-    events[event_id].signaled = true;
-    let entry = &events[event_id];
-    let notify_fn = if entry.event_type & EVT_NOTIFY_SIGNAL != 0 {
-        entry.notify_function.map(|f| (f, entry.notify_context))
-    } else {
-        None
-    };
-    Ok((event_id, notify_fn))
-}
-
-extern "efiapi" fn signal_event(event: efi::Event) -> Status {
-    log::debug!("BS.SignalEvent(event={:?})", event);
-
-    let notify = with_tables_mut(|efi_state| signal_event_entry(&mut efi_state.events, event));
-
-    let (event_id, notify_fn) = match notify {
-        Ok(result) => result,
-        Err(status) => return status,
-    };
-    if let Some((func, context)) = notify_fn {
-        log::debug!("  -> Calling notify function for event {}", event_id);
-        func(event, context);
-    }
-
-    Status::SUCCESS
-}
-
-fn close_dynamic_event(events: &mut [EventEntry], event: efi::Event) -> bool {
-    let Some(event_id) = dynamic_event_id_for_handle(events, event) else {
-        return false;
-    };
-    let generation = events[event_id].generation;
-    events[event_id] = EventEntry::empty();
-    events[event_id].generation = generation;
-    true
-}
-
-extern "efiapi" fn close_event(event: efi::Event) -> Status {
-    log::debug!("BS.CloseEvent(event={:?})", event);
-
-    with_tables_mut(|efi_state| {
-        if close_dynamic_event(&mut efi_state.events, event) {
-            // A registration outlives its event otherwise, and would keep
-            // signaling a closed event slot on every matching install.
-            efi_state
-                .protocol_notifies
-                .retain(|notify| notify.event != event);
-            Status::SUCCESS
-        } else {
-            Status::INVALID_PARAMETER
-        }
-    })
-}
-
-extern "efiapi" fn check_event(event: efi::Event) -> Status {
-    let Some(event_id) = with_tables_mut(|efi_state| event_id_for_handle(&efi_state.events, event))
-    else {
-        return Status::INVALID_PARAMETER;
-    };
-    log::trace!("BS.CheckEvent(event={:?}, slot={})", event, event_id);
-
-    // Special case for keyboard event — do a real read-ahead check
-    if event_id == KEYBOARD_EVENT_ID {
-        if crate::efi::protocols::console::keyboard_check_ready() {
-            return Status::SUCCESS;
-        } else {
-            return Status::NOT_READY;
-        }
-    }
-
-    // Special case for pointer event — poll hardware and peek
-    #[cfg(feature = "ui")]
-    if event_id == POINTER_EVENT_ID {
-        if crate::efi::protocols::simple_pointer::pointer_check_ready() {
-            return Status::SUCCESS;
-        } else {
-            return Status::NOT_READY;
-        }
-    }
-
-    // Check regular dynamic events.
-    if event_id >= FIRST_DYNAMIC_EVENT_ID {
-        notify_wait_event(event_id, event);
-
-        // Check timer expiration
-        check_timer_event(event_id);
-
-        // Per UEFI spec: CheckEvent clears the signaled state when
-        // returning SUCCESS (for EVT_NOTIFY_WAIT events, the notify
-        // function is called first, then the event is cleared).
-        let signaled = with_tables_mut(|efi_state| {
-            let was_signaled = efi_state.events[event_id].signaled;
-            if was_signaled {
-                efi_state.events[event_id].signaled = false;
-            }
-            was_signaled
-        });
-        if signaled {
-            return Status::SUCCESS;
-        }
-    }
-
-    Status::NOT_READY
-}
-
-/// Check if a timer event has expired and signal it if so
-fn check_timer_event(event_id: usize) {
-    with_tables_mut(|efi_state| {
-        let entry = &mut efi_state.events[event_id];
-
-        // Only process timer events with an active deadline
-        if entry.event_type & EVT_TIMER == 0 || entry.timer_type == TimerType::Cancel {
-            return;
-        }
-
-        if entry.timer_deadline_tsc == 0 {
-            return;
-        }
-
-        let now = crate::time::rdtsc();
-        if now >= entry.timer_deadline_tsc {
-            entry.signaled = true;
-
-            match entry.timer_type {
-                TimerType::Periodic => {
-                    // Reset deadline for next period
-                    let tsc_per_us = (crate::time::tsc_frequency() / 1_000_000).max(1);
-                    let us = entry.timer_trigger_time / 10;
-                    let tsc_offset = us * tsc_per_us;
-                    entry.timer_deadline_tsc = now + tsc_offset;
-                }
-                TimerType::Relative => {
-                    // One-shot: clear the deadline
-                    entry.timer_deadline_tsc = 0;
-                }
-                TimerType::Cancel => {}
-            }
-        }
-    });
-}
-
-/// EFI_EVENT_GROUP_READY_TO_BOOT GUID
-const EFI_EVENT_GROUP_READY_TO_BOOT: Guid = Guid::from_fields(
-    0x7CE88FB3,
-    0x4BD7,
-    0x4679,
-    0x87,
-    0xA8,
-    &[0xA8, 0xD8, 0xDE, 0xE5, 0x0D, 0x2B],
-);
-
-/// EFI_EVENT_GROUP_EXIT_BOOT_SERVICES GUID
-const EFI_EVENT_GROUP_EXIT_BOOT_SERVICES: Guid = Guid::from_fields(
-    0x27ABF055,
-    0xB1B8,
-    0x4C26,
-    0x80,
-    0x48,
-    &[0x74, 0x8F, 0x37, 0xBA, 0xA2, 0xDF],
-);
-
-/// Signal all events belonging to a specific event group
-fn signal_event_group(group_guid: &Guid) {
-    // Collect events to signal (must not hold a state borrow during callbacks)
-    let mut events_to_signal: heapless::Vec<
-        (usize, Option<efi::EventNotify>, *mut c_void),
-        MAX_EVENTS,
-    > = heapless::Vec::new();
-
-    with_tables_mut(|efi_state| {
-        for (i, entry) in efi_state.events.iter_mut().enumerate() {
-            if entry.in_use
-                && let Some(ref group) = entry.event_group
-                && *group == *group_guid
-            {
-                entry.signaled = true;
-                let notify = if entry.event_type & EVT_NOTIFY_SIGNAL != 0 {
-                    entry.notify_function.map(|f| (f, entry.notify_context))
-                } else {
-                    None
-                };
-                let _ = events_to_signal.push((
-                    event_handle(i, entry.generation) as usize,
-                    notify.map(|(f, _)| f),
-                    notify.map(|(_, c)| c).unwrap_or(core::ptr::null_mut()),
-                ));
-            }
-        }
-    });
-
-    // Call notify functions outside the state borrow
-    for (event_handle, notify_fn, context) in &events_to_signal {
-        if let Some(func) = notify_fn {
-            log::debug!(
-                "signal_event_group: calling notify for event {:#x}",
-                event_handle
-            );
-            func(*event_handle as efi::Event, *context);
-        }
-    }
-
-    if !events_to_signal.is_empty() {
-        log::info!(
-            "signal_event_group: signaled {} events",
-            events_to_signal.len()
-        );
-    }
-}
-
-/// Measure the beginning of an EFI boot application attempt.
-///
-/// The first attempt performs the ReadyToBoot measured-boot sequence and
-/// signals the ReadyToBoot event group. Subsequent attempts only add the boot
-/// action event so separators are not duplicated.
-pub(crate) fn measure_efi_application_start(is_application: bool) {
-    if !is_application {
-        return;
-    }
-
-    let should_signal = with_tables_mut(|efi| {
-        if !efi.ready_to_boot_signaled {
-            efi.ready_to_boot_signaled = true;
-            true
-        } else {
-            false
-        }
-    });
-
-    if should_signal {
-        // TCG measured boot: measure pre-OS handoff state, boot variables,
-        // boot action, and separators before the first boot attempt, per TCG
-        // PFP / EDK2 ReadyToBoot ordering.
-        super::tcg::measured_boot::measure_handoff_tables_all();
-        super::tcg::measured_boot::measure_boot_variables_all();
-        super::tcg::measured_boot::measure_action_all(
-            4,
-            "Calling EFI Application from Boot Option",
-        );
-
-        // Measure separator events into PCR 0-6.
-        // PCR 7 already has its separator from Secure Boot variable measurement.
-        super::tcg::measured_boot::measure_all_separators_all();
-
-        signal_event_group(&EFI_EVENT_GROUP_READY_TO_BOOT);
-    } else {
-        super::tcg::measured_boot::measure_action_all(
-            4,
-            "Calling EFI Application from Boot Option",
-        );
-    }
-}
-
-/// Measure return from an EFI boot application attempt.
-pub(crate) fn measure_efi_application_return(is_application: bool) {
-    if is_application && tables().ready_to_boot_signaled {
-        super::tcg::measured_boot::measure_action_all(
-            4,
-            "Returning from EFI Application from Boot Option",
-        );
-    }
-}
-
-extern "efiapi" fn create_event_ex(
-    event_type: u32,
-    notify_tpl: Tpl,
-    notify_function: Option<efi::EventNotify>,
-    notify_context: *const c_void,
-    event_group: *const Guid,
-    event: *mut efi::Event,
-) -> Status {
-    let group_display = if event_group.is_null() {
-        None
-    } else {
-        Some(GuidFmt(unsafe { *event_group }))
-    };
-    log::debug!(
-        "BS.CreateEventEx(type={:#x}, tpl={:?}, group={})",
-        event_type,
-        notify_tpl,
-        group_display
-            .as_ref()
-            .map(|g| g as &dyn core::fmt::Display)
-            .unwrap_or(&"NULL" as &dyn core::fmt::Display)
-    );
-
-    if event.is_null() {
-        return Status::INVALID_PARAMETER;
-    }
-
-    // Create the event with full parameters
-    let status = create_event(
-        event_type,
-        notify_tpl,
-        notify_function,
-        notify_context as *mut c_void,
-        event,
-    );
-
-    if status == Status::SUCCESS && !event_group.is_null() {
-        let handle = unsafe { *event };
-        with_tables_mut(|efi_state| {
-            if let Some(event_id) = dynamic_event_id_for_handle(&efi_state.events, handle) {
-                efi_state.events[event_id].event_group = Some(unsafe { *event_group });
-            }
-        });
-    }
-
-    status
-}
-
-// ============================================================================
-// Protocol Handler Functions
-// ============================================================================
-
-extern "efiapi" fn install_protocol_interface(
-    handle: *mut Handle,
-    protocol: *mut Guid,
-    interface_type: efi::InterfaceType,
-    interface: *mut c_void,
-) -> Status {
-    if handle.is_null() || protocol.is_null() {
-        return Status::INVALID_PARAMETER;
-    }
-
-    if interface_type != efi::NATIVE_INTERFACE {
-        return Status::INVALID_PARAMETER;
-    }
-
-    let target_handle = if unsafe { (*handle).is_null() } {
-        let Some(new_handle) = create_handle() else {
-            return Status::OUT_OF_RESOURCES;
-        };
-        unsafe { *handle = new_handle };
-        new_handle
-    } else {
-        unsafe { *handle }
-    };
-
-    install_protocol(target_handle, &unsafe { *protocol }, interface)
-}
-
-extern "efiapi" fn reinstall_protocol_interface(
-    handle: Handle,
-    protocol: *mut Guid,
-    old_interface: *mut c_void,
-    new_interface: *mut c_void,
-) -> Status {
-    if handle.is_null() || protocol.is_null() {
-        return Status::INVALID_PARAMETER;
-    }
-
-    let guid = unsafe { *protocol };
-    let (status, notify_events) = with_tables_mut(|efi_state| {
-        let Some(entry) = efi_state.handles[..efi_state.handle_count]
-            .iter_mut()
-            .find(|entry| entry.handle == handle)
-        else {
-            return (Status::NOT_FOUND, heapless::Vec::new());
-        };
-        let Some(protocol_entry) = entry.protocols[..entry.protocol_count]
-            .iter_mut()
-            .find(|entry| entry.guid == guid && entry.interface == old_interface)
-        else {
-            return (Status::NOT_FOUND, heapless::Vec::new());
-        };
-        if efi_state.open_protocols.iter().any(|open| {
-            open.handle == handle
-                && open.protocol == guid
-                && protocol_open_blocks_removal(open.attributes)
-        }) {
-            // Built-in drivers have no DisconnectController implementation. Do not
-            // replace an interface while a driver or exclusive user still holds it.
-            return (Status::ACCESS_DENIED, heapless::Vec::new());
-        }
-        efi_state
-            .open_protocols
-            .retain(|open| open.handle != handle || open.protocol != guid);
-        let Some(generation) = efi_state.protocol_generation.checked_add(1) else {
-            return (Status::OUT_OF_RESOURCES, heapless::Vec::new());
-        };
-        efi_state.protocol_generation = generation;
-        protocol_entry.generation = generation;
-        protocol_entry.interface = new_interface;
-        // Per the UEFI spec a reinstall notifies registrations just like an
-        // install, so drivers can rebind to the replacement interface.
-        let events = protocol_notification_events(efi_state, &guid);
-        (Status::SUCCESS, events)
-    });
-
-    for event in notify_events {
-        signal_event(event);
-    }
-
-    status
-}
-
-extern "efiapi" fn uninstall_protocol_interface(
-    handle: Handle,
-    protocol: *mut Guid,
-    interface: *mut c_void,
-) -> Status {
-    if handle.is_null() || protocol.is_null() {
-        return Status::INVALID_PARAMETER;
-    }
-    remove_protocol(handle, &unsafe { *protocol }, Some(interface))
-}
-
-extern "efiapi" fn handle_protocol(
-    handle: Handle,
-    protocol: *mut Guid,
-    interface: *mut *mut c_void,
-) -> Status {
-    let guid = if protocol.is_null() {
-        Guid::from_fields(0, 0, 0, 0, 0, &[0; 6])
-    } else {
-        unsafe { *protocol }
-    };
-    log::debug!(
-        "BS.HandleProtocol(handle={:?}, protocol={})",
-        handle,
-        GuidFmt(guid)
-    );
-
-    // Forward to open_protocol with simpler semantics
-    let status = open_protocol(
-        handle,
-        protocol,
-        interface,
-        core::ptr::null_mut(), // agent_handle
-        core::ptr::null_mut(), // controller_handle
-        efi::OPEN_PROTOCOL_BY_HANDLE_PROTOCOL,
-    );
-
-    if status != Status::SUCCESS {
-        log::debug!("  -> {:?}", status);
-    }
-
-    status
-}
-
-extern "efiapi" fn register_protocol_notify(
-    protocol: *mut Guid,
-    event: efi::Event,
-    registration: *mut *mut c_void,
-) -> Status {
-    if protocol.is_null() || registration.is_null() {
-        return Status::INVALID_PARAMETER;
-    }
-    let event_id = event as usize;
-    if event_id == 0 || event_id >= MAX_EVENTS {
-        return Status::INVALID_PARAMETER;
-    }
-
-    let guid = unsafe { *protocol };
-    log::debug!("BS.RegisterProtocolNotify(protocol={})", GuidFmt(guid));
-
-    with_tables_mut(|efi_state| {
-        if efi_state.protocol_notifies.len() >= MAX_PROTOCOL_NOTIFIES
-            || efi_state.protocol_notifies.try_reserve(1).is_err()
-        {
-            log::warn!("  -> OUT_OF_RESOURCES (notify registration table full)");
-            return Status::OUT_OF_RESOURCES;
-        }
-
-        let token = efi_state.next_registration;
-        let Some(next_token) = token.checked_add(1) else {
-            return Status::OUT_OF_RESOURCES;
-        };
-        efi_state.next_registration = next_token;
-        efi_state.protocol_notifies.push(ProtocolNotifyEntry {
-            registration: token,
-            protocol: guid,
-            event,
-            cursor: 0,
-        });
-
-        unsafe { *registration = token as *mut c_void };
-        log::debug!("  -> SUCCESS (registration={:#x})", token);
-        Status::SUCCESS
-    })
-}
-
-/// Collect events to signal after releasing the database borrow.
-fn protocol_notification_events(
-    efi_state: &Tables,
-    guid: &Guid,
-) -> heapless::Vec<efi::Event, MAX_PROTOCOL_NOTIFIES> {
-    efi_state
-        .protocol_notifies
-        .iter()
-        .filter(|notify| notify.protocol == *guid)
-        .map(|notify| notify.event)
-        .collect()
-}
-
-/// Peek the next live instance without advancing a registration's shared cursor.
-fn next_registered_protocol(
-    efi_state: &Tables,
-    token: usize,
-    protocol: Option<Guid>,
-) -> Result<(usize, Handle, ProtocolEntry), Status> {
-    let (index, notify) = efi_state
-        .protocol_notifies
-        .iter()
-        .enumerate()
-        .find(|(_, notify)| notify.registration == token)
-        .ok_or(Status::INVALID_PARAMETER)?;
-    if protocol.is_some_and(|guid| guid != notify.protocol) {
-        return Err(Status::INVALID_PARAMETER);
-    }
-    efi_state.handles[..efi_state.handle_count]
-        .iter()
-        .flat_map(|handle| {
-            handle.protocols[..handle.protocol_count]
-                .iter()
-                .map(move |entry| (index, handle.handle, *entry))
-        })
-        .filter(|(_, _, entry)| entry.guid == notify.protocol && entry.generation > notify.cursor)
-        .min_by_key(|(_, _, entry)| entry.generation)
-        .ok_or(Status::NOT_FOUND)
-}
-
-extern "efiapi" fn locate_handle(
-    search_type: efi::LocateSearchType,
-    protocol: *mut Guid,
-    search_key: *mut c_void,
-    buffer_size: *mut usize,
-    buffer: *mut Handle,
-) -> Status {
-    if buffer_size.is_null() {
-        return Status::INVALID_PARAMETER;
-    }
-
-    let guid_display = if protocol.is_null() {
-        None
-    } else {
-        Some(GuidFmt(unsafe { *protocol }))
-    };
-
-    log::debug!(
-        "BS.LocateHandle(type={}, protocol={}, buf_size={}, buf={:?})",
-        search_type,
-        guid_display
-            .as_ref()
-            .map(|g| g as &dyn core::fmt::Display)
-            .unwrap_or(&"NULL" as &dyn core::fmt::Display),
-        unsafe { *buffer_size },
-        buffer
-    );
-
-    if search_type == efi::BY_REGISTER_NOTIFY {
-        return with_tables_mut(|efi_state| {
-            let (index, handle, entry) =
-                match next_registered_protocol(efi_state, search_key as usize, None) {
-                    Ok(next) => next,
-                    Err(status) => return status,
-                };
-            let required_size = core::mem::size_of::<Handle>();
-            let capacity = unsafe { *buffer_size };
-            unsafe { *buffer_size = required_size };
-            if buffer.is_null() || capacity < required_size {
-                return Status::BUFFER_TOO_SMALL;
-            }
-            unsafe { *buffer = handle };
-            efi_state.protocol_notifies[index].cursor = entry.generation;
-            Status::SUCCESS
-        });
-    }
-
-    // Collect matching handles based on search type. The handle database grows
-    // past its preallocated size, so this collects onto the heap rather than
-    // into a fixed-capacity buffer that would silently truncate the result.
-    //
-    let matching = {
-        let efi_state = tables();
-        match search_type {
-            efi::ALL_HANDLES => match collect_handles(&efi_state, |_| true) {
-                Ok(handles) => handles,
-                Err(status) => return status,
-            },
-            efi::BY_PROTOCOL => {
-                if protocol.is_null() {
-                    return Status::INVALID_PARAMETER;
-                }
-                let guid = unsafe { *protocol };
-                match collect_handles(&efi_state, |entry| {
-                    entry.protocols[..entry.protocol_count]
-                        .iter()
-                        .any(|p| p.guid == guid)
-                }) {
-                    Ok(handles) => handles,
-                    Err(status) => return status,
-                }
-            }
-            _ => {
-                log::debug!(
-                    "  -> INVALID_PARAMETER (unknown search type {})",
-                    search_type
-                );
-                return Status::INVALID_PARAMETER;
-            }
-        }
-    };
-
-    // Check for no matches FIRST, before buffer size checks
-    if matching.is_empty() {
-        log::debug!("  -> NOT_FOUND (no matching handles)");
-        return Status::NOT_FOUND;
-    }
-
-    let required_size = matching.len() * core::mem::size_of::<Handle>();
-
-    if buffer.is_null() || unsafe { *buffer_size } < required_size {
-        unsafe { *buffer_size = required_size };
-        log::debug!("  -> BUFFER_TOO_SMALL (need {} bytes)", required_size);
-        return Status::BUFFER_TOO_SMALL;
-    }
-
-    // Copy handles to buffer using slice copy
-    let dest = unsafe { core::slice::from_raw_parts_mut(buffer, matching.len()) };
-    dest.copy_from_slice(&matching[..]);
-    unsafe { *buffer_size = required_size };
-
-    log::debug!("  -> found {} handles: {:?}", matching.len(), matching);
-
-    log::debug!("  -> found {} handles: {:?}", matching.len(), &matching[..]);
-    Status::SUCCESS
-}
-
-/// Collect the handles matching `filter` onto the firmware heap.
-///
-/// # Arguments
-/// * `efi_state` - Borrowed EFI state.
-/// * `filter` - Predicate applied to each live handle entry.
-///
-/// # Returns
-/// The matching handles, or `Err(OUT_OF_RESOURCES)` when the heap cannot hold them.
-fn collect_handles(
-    efi_state: &Tables,
-    filter: impl Fn(&HandleEntry) -> bool,
-) -> Result<Vec<Handle>, Status> {
-    let mut handles = Vec::new();
-    if handles.try_reserve_exact(efi_state.handle_count).is_err() {
-        return Err(Status::OUT_OF_RESOURCES);
-    }
-    handles.extend(
-        efi_state.handles[..efi_state.handle_count]
-            .iter()
-            .filter(|entry| filter(entry))
-            .map(|entry| entry.handle),
-    );
-    Ok(handles)
-}
-
-unsafe fn device_path_node_len(dp: *mut DevicePathProtocol) -> Option<usize> {
-    let len = unsafe { u16::from_le_bytes([(*dp).length[0], (*dp).length[1]]) as usize };
-    (len >= 4).then_some(len)
-}
-
-unsafe fn is_device_path_end(dp: *mut DevicePathProtocol) -> bool {
-    unsafe { (*dp).r#type == 0x7f && (*dp).sub_type == 0xff }
-}
-
-unsafe fn device_path_prefix_match(
-    handle_dp: *mut DevicePathProtocol,
-    input_dp: *mut DevicePathProtocol,
-) -> Option<*mut DevicePathProtocol> {
-    let mut handle_node = handle_dp;
-    let mut input_node = input_dp;
-
-    // Device paths are small. Bound the walk so malformed paths cannot loop forever.
-    for _ in 0..128 {
-        if unsafe { is_device_path_end(handle_node) } {
-            return Some(input_node);
-        }
-        if unsafe { is_device_path_end(input_node) } {
-            return None;
-        }
-
-        let handle_len = unsafe { device_path_node_len(handle_node)? };
-        let input_len = unsafe { device_path_node_len(input_node)? };
-        if handle_len != input_len {
-            return None;
-        }
-
-        let handle_bytes =
-            unsafe { core::slice::from_raw_parts(handle_node as *const u8, handle_len) };
-        let input_bytes =
-            unsafe { core::slice::from_raw_parts(input_node as *const u8, input_len) };
-        if handle_bytes != input_bytes {
-            return None;
-        }
-
-        handle_node =
-            unsafe { (handle_node as *const u8).add(handle_len) as *mut DevicePathProtocol };
-        input_node = unsafe { (input_node as *const u8).add(input_len) as *mut DevicePathProtocol };
-    }
-
-    None
-}
-
-extern "efiapi" fn locate_device_path(
-    protocol: *mut Guid,
-    device_path: *mut *mut DevicePathProtocol,
-    device: *mut Handle,
-) -> Status {
-    if protocol.is_null() || device_path.is_null() || device.is_null() {
-        return Status::INVALID_PARAMETER;
-    }
-
-    let guid = unsafe { *protocol };
-    log::debug!("BS.LocateDevicePath(protocol={})", GuidFmt(guid));
-
-    let input_dp = unsafe { *device_path };
-    if input_dp.is_null() {
-        log::debug!("  -> INVALID_PARAMETER (device_path is NULL)");
-        return Status::INVALID_PARAMETER;
-    }
-
-    // Find a handle with both the specified protocol and a DEVICE_PATH protocol
-    let efi_state = tables();
-
-    let found = efi_state.handles[..efi_state.handle_count]
-        .iter()
-        .find_map(|entry| {
-            let protocols = &entry.protocols[..entry.protocol_count];
-
-            let has_protocol = protocols.iter().any(|p| p.guid == guid);
-            if !has_protocol {
-                return None;
-            }
-
-            let handle_dp = protocols
-                .iter()
-                .find(|p| p.guid == r_efi::protocols::device_path::PROTOCOL_GUID)
-                .map(|p| p.interface as *mut DevicePathProtocol)?;
-            if handle_dp.is_null() {
-                return None;
-            }
-
-            let remaining = unsafe { device_path_prefix_match(handle_dp, input_dp) }?;
-            Some((entry.handle, remaining))
-        });
-
-    if let Some((handle, remaining)) = found {
-        log::debug!(
-            "  -> SUCCESS (handle={:?}, remaining_device_path={:?})",
-            handle,
-            remaining
-        );
-        unsafe {
-            *device = handle;
-            *device_path = remaining;
-        }
-        return Status::SUCCESS;
-    }
-
-    log::debug!("  -> NOT_FOUND");
-    Status::NOT_FOUND
-}
-
-extern "efiapi" fn install_configuration_table(guid: *mut Guid, table: *mut c_void) -> Status {
-    if guid.is_null() {
-        return Status::INVALID_PARAMETER;
-    }
-
-    let guid_ref = unsafe { &*guid };
-    system_table::install_configuration_table(guid_ref, table)
-}
-
-// ============================================================================
-// Image Functions
-// ============================================================================
-// Device path parsing and file loading helpers are in `super::image_loader`.
-
-extern "efiapi" fn load_image(
-    boot_policy: Boolean,
-    parent_image_handle: Handle,
-    device_path: *mut DevicePathProtocol,
-    source_buffer: *mut c_void,
-    source_size: usize,
-    image_handle: *mut Handle,
-) -> Status {
-    log::debug!(
-        "BS.LoadImage(boot_policy={:?}, parent={:?}, device_path={:?}, buf={:?}, size={})",
-        boot_policy,
-        parent_image_handle,
-        device_path,
-        source_buffer,
-        source_size
-    );
-
-    // Validate parameters
-    if image_handle.is_null() {
-        log::error!("BS.LoadImage: image_handle is NULL");
-        return Status::INVALID_PARAMETER;
-    }
-
-    // Determine the image source: either a caller-provided buffer or loaded from device path.
-    enum ImageSource {
-        /// Caller-provided buffer — not owned by us, must not be freed
-        Buffer {
-            data_ptr: *mut c_void,
-            data_size: usize,
-        },
-        /// Loaded from device path — we allocated this buffer and must free it
-        DevicePath {
-            data_ptr: *mut c_void,
-            data_size: usize,
-            device_handle: Handle,
-        },
-    }
-
-    let source = if !source_buffer.is_null() && source_size > 0 {
-        ImageSource::Buffer {
-            data_ptr: source_buffer,
-            data_size: source_size,
-        }
-    } else if !device_path.is_null() {
-        match image_loader::load_image_from_device_path(device_path) {
-            Ok((ptr, size, dev_handle)) => ImageSource::DevicePath {
-                data_ptr: ptr,
-                data_size: size,
-                device_handle: dev_handle,
-            },
-            Err(status) => {
-                log::error!(
-                    "BS.LoadImage: Failed to load from device path: {:?}",
-                    status
-                );
-                return status;
-            }
-        }
-    } else {
-        log::error!("BS.LoadImage: No source buffer and no device path provided");
-        return Status::INVALID_PARAMETER;
-    };
-
-    let (data_ptr, data_size) = match &source {
-        ImageSource::Buffer {
-            data_ptr,
-            data_size,
-        } => (*data_ptr, *data_size),
-        ImageSource::DevicePath {
-            data_ptr,
-            data_size,
-            ..
-        } => (*data_ptr, *data_size),
-    };
-
-    // Create a slice from the source buffer
-    let data = unsafe { core::slice::from_raw_parts(data_ptr as *const u8, data_size) };
-
-    // Helper to free the buffer only if we own it (loaded from device path)
-    let free_if_owned = |source: &ImageSource| {
-        if let ImageSource::DevicePath { data_ptr, .. } = source {
-            let _ = allocator::free_pool(*data_ptr as *mut u8);
-        }
-    };
-
-    // Secure Boot verification (if enabled)
-    if super::auth::is_secure_boot_enabled() {
-        log::debug!("BS.LoadImage: Secure Boot verification required");
-        match super::auth::verify_pe_image_secure_boot(data) {
-            Ok(true) => {
-                log::info!("BS.LoadImage: Secure Boot verification passed");
-            }
-            Ok(false) => {
-                log::error!("BS.LoadImage: Secure Boot verification FAILED - image not authorized");
-                crate::display_secure_boot_error();
-                free_if_owned(&source);
-                return Status::SECURITY_VIOLATION;
-            }
-            Err(e) => {
-                log::error!("BS.LoadImage: Secure Boot verification error: {:?}", e);
-                crate::display_secure_boot_error();
-                free_if_owned(&source);
-                return Status::SECURITY_VIOLATION;
-            }
-        }
-    }
-
-    // Load the PE image using our PE loader
-    let loaded_image = match pe::load_image(data) {
-        Ok(img) => img,
-        Err(status) => {
-            log::error!("BS.LoadImage: Failed to load PE image: {:?}", status);
-            free_if_owned(&source);
-            return status;
-        }
-    };
-
-    log::debug!(
-        "BS.LoadImage: PE loaded at {:#x}, entry={:#x}, size={:#x}",
-        loaded_image.image_base,
-        loaded_image.entry_point,
-        loaded_image.image_size
-    );
-
-    // Preserve metadata before releasing an owned source buffer.
-    let image_subsystem = pe::parse_headers(data)
-        .map(|headers| headers.subsystem())
-        .unwrap_or(0);
-
-    // TCG measured boot: drivers are measured now; applications are deferred
-    // until StartImage after ReadyToBoot, with digests computed before freeing data.
-    let deferred_measurement = measure_pe_image_for_tcg(
-        data,
-        &loaded_image,
-        device_path as *const DevicePathProtocol,
-    );
-
-    // Free the buffer now that PE is loaded and measured (PE loading makes its own copy).
-    free_if_owned(&source);
-
-    // Create a new handle for this image
-    let new_handle = match create_handle() {
-        Some(h) => h,
-        None => {
-            log::error!("BS.LoadImage: Failed to create handle");
-            pe::unload_image(&loaded_image);
-            return Status::OUT_OF_RESOURCES;
-        }
-    };
-
-    // Create LoadedImageProtocol for this image
-    // Use the device handle from loading if we loaded from device path,
-    // otherwise try to get it from the parent
-    let device_handle = match &source {
-        ImageSource::DevicePath { device_handle, .. } => *device_handle,
-        _ => image_loader::get_device_handle_from_parent(parent_image_handle),
-    };
-
-    let system_table = super::get_system_table();
-    let loaded_image_protocol = create_loaded_image_protocol(
-        parent_image_handle,
-        system_table,
-        device_handle,
-        loaded_image.image_base,
-        loaded_image.image_size,
-    );
-
-    if loaded_image_protocol.is_null() {
-        log::error!("BS.LoadImage: Failed to create LoadedImageProtocol");
-        if let Some(measurement) = deferred_measurement {
-            let _ = allocator::free_pool(measurement.event_data);
-        }
-        pe::unload_image(&loaded_image);
-        return Status::OUT_OF_RESOURCES;
-    }
-
-    // Set the device path on the loaded image if provided
-    if !device_path.is_null() {
-        unsafe {
-            super::protocols::loaded_image::set_file_path(loaded_image_protocol, device_path);
-        }
-    }
-
-    // Install the LoadedImageProtocol on the handle
-    let status = install_protocol(
-        new_handle,
-        &LOADED_IMAGE_PROTOCOL_GUID,
-        loaded_image_protocol as *mut c_void,
-    );
-
-    if status != Status::SUCCESS {
-        log::error!(
-            "BS.LoadImage: Failed to install LoadedImageProtocol: {:?}",
-            status
-        );
-        if let Some(measurement) = deferred_measurement {
-            let _ = allocator::free_pool(measurement.event_data);
-        }
-        pe::unload_image(&loaded_image);
-        return status;
-    }
-
-    // Store the loaded image info so StartImage can find it
-    let store_result = with_tables_mut(|efi_state| {
-        let slot = efi_state
-            .loaded_images
-            .iter_mut()
-            .find(|entry| entry.handle.is_null());
-
-        match slot {
-            Some(entry) => {
-                entry.handle = new_handle;
-                entry.image_base = loaded_image.image_base;
-                entry.image_size = loaded_image.image_size;
-                entry.entry_point = loaded_image.entry_point;
-                entry.alloc_base = loaded_image.alloc_base;
-                entry.num_pages = loaded_image.num_pages;
-                entry.parent_handle = parent_image_handle;
-                entry.subsystem = image_subsystem;
-                if let Some(measurement) = deferred_measurement {
-                    entry.measurement_pcr = measurement.pcr_index;
-                    entry.measurement_event_type = measurement.event_type;
-                    entry.measurement_digest_count = measurement.digest_count;
-                    entry.measurement_digests = measurement.digests;
-                    entry.measurement_event_data = measurement.event_data;
-                    entry.measurement_event_data_size = measurement.event_data_size;
-                }
-                true
-            }
-            None => false,
-        }
-    });
-
-    if !store_result {
-        log::error!("BS.LoadImage: No space in loaded images table");
-        if let Some(measurement) = deferred_measurement {
-            let _ = allocator::free_pool(measurement.event_data);
-        }
-        pe::unload_image(&loaded_image);
-        return Status::OUT_OF_RESOURCES;
-    }
-
-    // Return the new handle
-    unsafe {
-        *image_handle = new_handle;
-    }
-
-    log::info!(
-        "BS.LoadImage: SUCCESS - handle={:?}, base={:#x}, entry={:#x}",
-        new_handle,
-        loaded_image.image_base,
-        loaded_image.entry_point
-    );
-
-    Status::SUCCESS
-}
-
-extern "efiapi" fn start_image(
-    image_handle: Handle,
-    exit_data_size: *mut usize,
-    exit_data: *mut *mut u16,
-) -> Status {
-    log::debug!("BS.StartImage(handle={:?})", image_handle);
-
-    if image_handle.is_null() {
-        log::error!("BS.StartImage: image_handle is NULL");
-        return Status::INVALID_PARAMETER;
-    }
-
-    // Find the loaded image entry
-    let (entry_point, image_base, image_subsystem) = {
-        let efi_state = tables();
-        match efi_state
-            .loaded_images
-            .iter()
-            .find(|entry| entry.handle == image_handle)
-            .map(|entry| (entry.entry_point, entry.image_base, entry.subsystem))
-        {
-            Some(info) => info,
-            None => {
-                log::error!(
-                    "BS.StartImage: handle {:?} not found in loaded images",
-                    image_handle
-                );
-                return Status::INVALID_PARAMETER;
-            }
-        }
-    };
-
-    log::info!(
-        "BS.StartImage: Executing image at {:#x} (base={:#x})",
-        entry_point,
-        image_base
-    );
-
-    // Signal EFI_EVENT_GROUP_READY_TO_BOOT before the first image is started
-    // and measure boot-attempt action events without duplicating separators.
-    let is_application = image_subsystem == 10;
-    measure_efi_application_start(is_application);
-
-    let deferred_measurement = with_tables_mut(|efi_state| {
-        efi_state
-            .loaded_images
-            .iter_mut()
-            .find(|entry| entry.handle == image_handle)
-            .and_then(|entry| {
-                if entry.measurement_event_data.is_null() {
-                    return None;
-                }
-                let measurement = DeferredImageMeasurement {
-                    pcr_index: entry.measurement_pcr,
-                    event_type: entry.measurement_event_type,
-                    digest_count: entry.measurement_digest_count,
-                    digests: entry.measurement_digests,
-                    event_data: entry.measurement_event_data,
-                    event_data_size: entry.measurement_event_data_size,
-                };
-                entry.measurement_event_data = core::ptr::null_mut();
-                entry.measurement_event_data_size = 0;
-                entry.measurement_digest_count = 0;
-                Some(measurement)
-            })
-    });
-
-    if let Some(measurement) = deferred_measurement {
-        let event_data = unsafe {
-            // SAFETY: deferred measurement event data was allocated and filled in
-            // LoadImage and remains owned by this loaded-image entry until now.
-            core::slice::from_raw_parts(measurement.event_data, measurement.event_data_size)
-        };
-        if let Err(e) = super::tcg::measured_boot::measure_pe_image_digests_all(
-            measurement.pcr_index,
-            measurement.event_type,
-            &measurement.digests[..measurement.digest_count],
-            event_data,
-        ) {
-            log::warn!("Failed to measure PE image: {:?}", e);
-        }
-        let _ = allocator::free_pool(measurement.event_data);
-    }
-
-    // Update table CRC32s one final time before handing off to the image
-    // (config tables may have changed since efi::init())
-    super::system_table::update_crc32();
-
-    // Get the system table
-    let system_table = super::get_system_table();
-
-    // Define the entry point function type
-    type EfiEntryPoint = extern "efiapi" fn(Handle, *mut SystemTable) -> Status;
-
-    // Call the entry point
-    let entry: EfiEntryPoint = unsafe { core::mem::transmute(entry_point) };
-    let status = entry(image_handle, system_table);
-
-    log::info!("BS.StartImage: Image returned with status: {:?}", status);
-    measure_efi_application_return(is_application);
-
-    // Set exit data if provided (we don't support exit data currently)
-    if !exit_data_size.is_null() {
-        unsafe {
-            *exit_data_size = 0;
-        }
-    }
-    if !exit_data.is_null() {
-        unsafe {
-            *exit_data = core::ptr::null_mut();
-        }
-    }
-
-    status
-}
-
-/// EFI Boot Service: Exit
-///
-/// UEFI Spec Compliance Note: A fully conformant `Exit()` implementation must
-/// perform a non-local return (longjmp) back to the corresponding `StartImage()`
-/// call, unwinding the call stack. This requires saving the execution context
-/// (registers, stack pointer) in `StartImage()` via setjmp, and restoring it here.
-///
-/// Current limitation: This implementation simply returns `exit_status` to the
-/// caller, which means `Exit()` only works correctly when called directly from
-/// the image's entry point (the common case for UEFI bootloaders like shim and
-/// GRUB). It will NOT correctly unwind nested image calls or calls from deep
-/// within a loaded image's call stack.
-///
-/// This is acceptable for our boot use case (shim → GRUB → Linux), but would
-/// need a proper setjmp/longjmp implementation for full UEFI application support.
-extern "efiapi" fn exit(
-    image_handle: Handle,
-    exit_status: Status,
-    exit_data_size: usize,
-    _exit_data: *mut u16,
-) -> Status {
-    log::info!(
-        "BS.Exit(handle={:?}, status={:?}, data_size={})",
-        image_handle,
-        exit_status,
-        exit_data_size
-    );
-    exit_status
-}
-
-extern "efiapi" fn unload_image(image_handle: Handle) -> Status {
-    log::debug!("BS.UnloadImage(handle={:?})", image_handle);
-
-    if image_handle.is_null() {
-        log::error!("BS.UnloadImage: image_handle is NULL");
-        return Status::INVALID_PARAMETER;
-    }
-
-    // Refuse to free an image whose handle is still managed by another agent.
-    if tables().open_protocols.iter().any(|open| {
-        (open.handle == image_handle || open.controller_handle == image_handle)
-            && open.agent_handle != image_handle
-            && protocol_open_blocks_removal(open.attributes)
-    }) {
-        return Status::ACCESS_DENIED;
-    }
-
-    // Find and remove the loaded image entry
-    let image_info = with_tables_mut(|efi_state| {
-        efi_state
-            .loaded_images
-            .iter_mut()
-            .find(|entry| entry.handle == image_handle)
-            .map(|entry| {
-                let result = (
-                    entry.alloc_base,
-                    entry.num_pages,
-                    entry.measurement_event_data,
-                );
-                // Clear the entry
-                *entry = LoadedImageEntry::empty();
-                result
-            })
-    });
-
-    match image_info {
-        Some((alloc_base, num_pages, measurement_event_data)) => {
-            // Drop image-owned relationships before its code and data disappear.
-            with_tables_mut(|efi_state| {
-                efi_state.open_protocols.retain(|open| {
-                    open.agent_handle != image_handle
-                        && open.controller_handle != image_handle
-                        && open.handle != image_handle
-                });
-            });
-            // Free the image memory (using alloc_base, not image_base,
-            // since the image may have been aligned within the allocation)
-            let status = allocator::free_pages(alloc_base, num_pages);
-            if status != Status::SUCCESS {
-                log::warn!(
-                    "BS.UnloadImage: Failed to free pages at {:#x}: {:?}",
-                    alloc_base,
-                    status
-                );
-            }
-
-            if !measurement_event_data.is_null() {
-                let _ = allocator::free_pool(measurement_event_data);
-            }
-
-            // Remove protocols from the handle
-            // Note: In a full implementation, we should uninstall all protocols
-            // For now, we just log success
-            log::debug!("BS.UnloadImage: SUCCESS");
-            Status::SUCCESS
-        }
-        None => {
-            log::warn!(
-                "BS.UnloadImage: handle {:?} not found in loaded images",
-                image_handle
-            );
-            // Return success anyway - the handle might have been loaded differently
-            Status::SUCCESS
-        }
-    }
-}
-
-extern "efiapi" fn exit_boot_services(image_handle: Handle, map_key: usize) -> Status {
-    log::info!(
-        "BS.ExitBootServices(handle={:?}, map_key={:#x})",
-        image_handle,
-        map_key
-    );
-
-    // Reject a stale key before callbacks, measurements, or any irreversible
-    // transition. The allocator repeats this check at the actual commit point
-    // in case an EBS callback changes the map.
-    let key_status = allocator::validate_map_key(map_key);
-    if key_status != Status::SUCCESS {
-        return key_status;
-    }
-    let Some(runtime_image) = crate::efi::runtime_image::installed() else {
-        log::error!("ExitBootServices refused: runtime image client is missing");
-        return Status::DEVICE_ERROR;
-    };
-
-    // TCG measured boot: measure ExitBootServices action into PCR 5.
-    super::tcg::measured_boot::measure_action_all(5, "Exit Boot Services Invocation");
-
-    // Signal EXIT_BOOT_SERVICES event group BEFORE finalizing the memory map.
-    // Windows Boot Manager registers callbacks that must run before we lock
-    // the memory map.
-    signal_event_group(&EFI_EVENT_GROUP_EXIT_BOOT_SERVICES);
-
-    // Also signal any legacy EVT_SIGNAL_EXIT_BOOT_SERVICES events
-    {
-        let mut legacy_events: heapless::Vec<usize, MAX_EVENTS> = heapless::Vec::new();
-        with_tables_mut(|efi_state| {
-            for (i, event) in efi_state.events.iter_mut().enumerate() {
-                if event.in_use && event.event_type == EVT_SIGNAL_EXIT_BOOT_SERVICES {
-                    event.signaled = true;
-                    let _ = legacy_events.push(event_handle(i, event.generation) as usize);
-                }
-            }
-        });
-        for handle in &legacy_events {
-            let notify_fn = {
-                let efi_state = tables();
-                dynamic_event_id_for_handle(&efi_state.events, *handle as efi::Event).and_then(
-                    |event_id| {
-                        let entry = &efi_state.events[event_id];
-                        entry.notify_function.map(|f| (f, entry.notify_context))
-                    },
-                )
-            };
-            if let Some((func, context)) = notify_fn {
-                func(*handle as efi::Event, context);
-            }
-        }
-    }
-
-    // Rebuild the Memory Attributes Table in-place BEFORE locking the allocator.
-    // Runtime image and retained-buffer regions are registered after the
-    // initial table setup, so the final MAT must be rebuilt from the allocator.
-    // A stale MEMATTR table with missing entries causes Windows to crash.
-    // We use the in-place variant that overwrites the existing page without
-    // calling allocate_pages(), so the map_key stays valid for the caller.
-    let prepare_status = system_table::rebuild_memory_attributes_table_in_place();
-    if prepare_status != Status::SUCCESS {
-        return prepare_status;
-    }
-
-    // Event callbacks and MAT rebuilding may have changed the memory map. Do a
-    // final key check before any irreversible hardware quiescence.
-    let key_status = allocator::validate_map_key(map_key);
-    if key_status != Status::SUCCESS {
-        return key_status;
-    }
-
-    // Stop every firmware-owned DMA engine while BootServices allocations are
-    // still typed and cannot yet be reused by the OS. Clearing BME is the final
-    // safety net for devices without complete driver shutdown coverage.
-    crate::drivers::quiesce_dma_for_os_handoff();
-
-    let status = allocator::exit_boot_services(map_key);
-
-    if status == Status::SUCCESS {
-        // TCG measured boot: measure ExitBootServices success into PCR 5.
-        super::tcg::measured_boot::measure_action_all(
-            5,
-            "Exit Boot Services Returned with Success",
-        );
-
-        log::info!("ExitBootServices SUCCESS - transitioning to OS");
-        crate::timestamp::record(crate::timestamp::TS_CRABEFI_EXIT_BOOT_SERVICES);
-
-        // Clean up hardware state for OS handoff.
-        // Re-enable keyboard interrupts so Linux's i8042 driver works.
-        crate::drivers::keyboard_common::cleanup();
-
-        // Seal only after the allocator accepted the map key, while boot-time
-        // diagnostics are still reachable. A failed seal leaves no safe way to
-        // return to the OS after allocator EBS, so report it and halt explicitly.
-        if let Err(seal_status) = runtime_image.seal() {
-            log::error!(
-                "FATAL: runtime image seal failed after allocator ExitBootServices: {:?}",
-                seal_status
-            );
-            loop {
-                crate::arch::halt();
-            }
-        }
-        log::info!("Runtime image sealed successfully");
-
-        // Let platform glue clean up integration-specific handoff state only
-        // after the final fallible step; hooks may disable non-runtime log
-        // buffers needed to diagnose a seal failure.
-        if let Some(hooks) = crate::handoff::callbacks().hooks {
-            hooks.on_exit_boot_services();
-        }
-
-        // CRITICAL: Disable logging only after the final fallible runtime-image
-        // transition. The OS generally does not map firmware log devices as
-        // runtime memory.
-        log::set_max_level(log::LevelFilter::Off);
-
-        // Switch from Secure EL1 to Non-Secure EL1 via a RAM trampoline.
-        //
-        // At Secure EL1, GICv3 routes Non-Secure Group 1 interrupts (LPIs /
-        // MSI-X) as FIQ. The Linux kernel only handles IRQ, so NVMe and other
-        // MSI-X devices hang forever waiting for completion interrupts.
-        //
-        // We can't issue the SMC directly from flash because the ERET returns
-        // to the instruction after the SMC — which is in Secure flash, not
-        // accessible from NS-EL1 on QEMU virt. Instead, we write a small
-        // trampoline to RAM that does SMC + RET. The RET returns to the EFI
-        // stub (also in RAM), now at NS-EL1 with proper interrupt routing.
-        //
-        // Uses vendor-specific SMCCC function ID 0xC2000000 handled by
-        // fstart's EL3 exception vector. No-op if no EL3 exists.
-        #[cfg(target_arch = "aarch64")]
-        crate::arch::aarch64::ns_switch::install_ns_trampoline();
-    } else {
-        super::tcg::measured_boot::measure_action_all(
-            5,
-            "Exit Boot Services Returned with Failure",
-        );
-        log::warn!("ExitBootServices FAILED: {:?}", status);
-    }
-
-    status
-}
-
-// ============================================================================
-// Miscellaneous Functions
-// ============================================================================
-
-extern "efiapi" fn get_next_monotonic_count(count: *mut u64) -> Status {
-    if count.is_null() {
-        return Status::INVALID_PARAMETER;
-    }
-
-    with_tables_mut(|efi_state| {
-        efi_state.monotonic_count += 1;
-        unsafe { *count = efi_state.monotonic_count };
-        Status::SUCCESS
-    })
-}
-
-extern "efiapi" fn stall(microseconds: usize) -> Status {
-    log::debug!("BS.Stall({}us)", microseconds);
-    // Use TSC-calibrated delay for accurate timing
-    crate::time::delay_us(microseconds as u64);
-    Status::SUCCESS
-}
-
-extern "efiapi" fn set_watchdog_timer(
-    timeout: usize,
-    watchdog_code: u64,
-    _data_size: usize,
-    _watchdog_data: *mut u16,
-) -> Status {
-    log::debug!(
-        "BS.SetWatchdogTimer(timeout={}, code={:#x})",
-        timeout,
-        watchdog_code
-    );
-    // Accept the call but don't implement actual watchdog.
-    // The UEFI spec default is a 5-minute watchdog that bootloaders disable
-    // by calling SetWatchdogTimer(0, 0, 0, NULL). Returning SUCCESS lets
-    // Windows Boot Manager proceed without error.
-    Status::SUCCESS
-}
-
-extern "efiapi" fn connect_controller(
-    controller_handle: Handle,
-    _driver_image_handle: *mut Handle,
-    _remaining_device_path: *mut DevicePathProtocol,
-    _recursive: Boolean,
-) -> Status {
-    log::debug!("BS.ConnectController(handle={:?})", controller_handle);
-    // CrabEFI doesn't use the UEFI driver model -- all drivers are built-in.
-    // Return SUCCESS so callers (like Windows Boot Manager) don't fail.
-    if controller_handle.is_null() {
-        return Status::INVALID_PARAMETER;
-    }
-    Status::SUCCESS
-}
-
-extern "efiapi" fn disconnect_controller(
-    controller_handle: Handle,
-    _driver_image_handle: Handle,
-    _child_handle: Handle,
-) -> Status {
-    log::debug!("BS.DisconnectController(handle={:?})", controller_handle);
-    // No-op for the same reason as ConnectController.
-    if controller_handle.is_null() {
-        return Status::INVALID_PARAMETER;
-    }
-    Status::SUCCESS
-}
-
-extern "efiapi" fn open_protocol(
-    handle: Handle,
-    protocol: *mut Guid,
-    interface: *mut *mut c_void,
-    agent_handle: Handle,
-    controller_handle: Handle,
-    attributes: u32,
-) -> Status {
-    if handle.is_null() || protocol.is_null() {
-        return Status::INVALID_PARAMETER;
-    }
-
-    const DRIVER_EXCLUSIVE: u32 = efi::OPEN_PROTOCOL_BY_DRIVER | efi::OPEN_PROTOCOL_EXCLUSIVE;
-    let base_attributes = attributes & !efi::OPEN_PROTOCOL_EXCLUSIVE;
-    if !matches!(
-        attributes,
-        efi::OPEN_PROTOCOL_BY_HANDLE_PROTOCOL
-            | efi::OPEN_PROTOCOL_GET_PROTOCOL
-            | efi::OPEN_PROTOCOL_TEST_PROTOCOL
-            | efi::OPEN_PROTOCOL_BY_CHILD_CONTROLLER
-            | efi::OPEN_PROTOCOL_BY_DRIVER
-            | efi::OPEN_PROTOCOL_EXCLUSIVE
-            | DRIVER_EXCLUSIVE
-    ) || ((matches!(
-        base_attributes,
-        efi::OPEN_PROTOCOL_BY_CHILD_CONTROLLER | efi::OPEN_PROTOCOL_BY_DRIVER
-    ) || attributes & efi::OPEN_PROTOCOL_EXCLUSIVE != 0)
-        && agent_handle.is_null())
-        || (attributes != efi::OPEN_PROTOCOL_TEST_PROTOCOL && interface.is_null())
-        || (matches!(
-            base_attributes,
-            efi::OPEN_PROTOCOL_BY_CHILD_CONTROLLER | efi::OPEN_PROTOCOL_BY_DRIVER
-        ) && controller_handle.is_null())
-        || (base_attributes == efi::OPEN_PROTOCOL_BY_CHILD_CONTROLLER
-            && controller_handle == handle)
-    {
-        return Status::INVALID_PARAMETER;
-    }
-
-    let guid = unsafe { *protocol };
-    log::debug!(
-        "BS.OpenProtocol(handle={:?}, protocol={}, attr={:#x})",
-        handle,
-        GuidFmt(guid),
-        attributes
-    );
-
-    let result = with_tables_mut(|efi_state| {
-        let Some(handle_entry) = efi_state.handles[..efi_state.handle_count]
-            .iter()
-            .find(|entry| entry.handle == handle)
-        else {
-            return Err(Status::INVALID_PARAMETER);
-        };
-        let Some(protocol_entry) = handle_entry.protocols[..handle_entry.protocol_count]
-            .iter()
-            .find(|entry| entry.guid == guid)
-        else {
-            return Err(Status::UNSUPPORTED);
-        };
-        let protocol_interface = protocol_entry.interface;
-        let exclusive = attributes & efi::OPEN_PROTOCOL_EXCLUSIVE != 0;
-
-        // Ordinary opens with an agent are observable through OpenProtocolInformation
-        // and CloseProtocol, but do not prevent interface teardown.
-        if agent_handle.is_null() {
-            return Ok(protocol_interface);
-        }
-
-        // A repeated driver open returns its existing interface without adding
-        // another relationship, including BY_DRIVER | EXCLUSIVE.
-        if attributes & efi::OPEN_PROTOCOL_BY_DRIVER != 0
-            && efi_state.open_protocols.iter().any(|open| {
-                open.handle == handle
-                    && open.protocol == guid
-                    && open.agent_handle == agent_handle
-                    && open.controller_handle == controller_handle
-                    && open.attributes == attributes
-            })
-        {
-            unsafe { *interface = protocol_interface };
-            return Err(Status::ALREADY_STARTED);
-        }
-
-        // There is no driver-disconnect implementation to evict an existing
-        // owner. Reject rather than pretending to grant exclusive access.
-        if (exclusive || base_attributes == efi::OPEN_PROTOCOL_BY_DRIVER)
-            && efi_state.open_protocols.iter().any(|open| {
-                open.handle == handle
-                    && open.protocol == guid
-                    && open.attributes
-                        & (efi::OPEN_PROTOCOL_BY_DRIVER | efi::OPEN_PROTOCOL_EXCLUSIVE)
-                        != 0
-            })
-        {
-            return Err(Status::ACCESS_DENIED);
-        }
-
-        if let Some(open) = efi_state.open_protocols.iter_mut().find(|open| {
-            open.handle == handle
-                && open.protocol == guid
-                && open.agent_handle == agent_handle
-                && open.controller_handle == controller_handle
-                && open.attributes == attributes
-        }) {
-            open.open_count = open.open_count.saturating_add(1);
-        } else {
-            if efi_state.open_protocols.try_reserve(1).is_err() {
-                return Err(Status::OUT_OF_RESOURCES);
-            }
-            efi_state.open_protocols.push(OpenProtocolEntry {
-                handle,
-                protocol: guid,
-                agent_handle,
-                controller_handle,
-                attributes,
-                open_count: 1,
-            });
-        }
-
-        Ok(protocol_interface)
-    });
-
-    let iface = match result {
-        Ok(iface) => iface,
-        Err(status) => return status,
-    };
-    if attributes != efi::OPEN_PROTOCOL_TEST_PROTOCOL && !interface.is_null() {
-        unsafe { *interface = iface };
-    }
-
-    if guid == LOADED_IMAGE_PROTOCOL_GUID && !iface.is_null() {
-        let lip = iface as *const r_efi::protocols::loaded_image::Protocol;
-        log::trace!(
-            "  -> LOADED_IMAGE(DeviceHandle={:?}, SystemTable={:?})",
-            unsafe { (*lip).device_handle },
-            unsafe { (*lip).system_table }
-        );
-    }
-
-    Status::SUCCESS
-}
-
-extern "efiapi" fn close_protocol(
-    handle: Handle,
-    protocol: *mut Guid,
-    agent_handle: Handle,
-    controller_handle: Handle,
-) -> Status {
-    if handle.is_null() || protocol.is_null() || agent_handle.is_null() {
-        return Status::INVALID_PARAMETER;
-    }
-    let guid = unsafe { *protocol };
-
-    with_tables_mut(|efi_state| {
-        let previous_len = efi_state.open_protocols.len();
-        // CloseProtocol closes all matching records, regardless of OpenCount.
-        efi_state.open_protocols.retain(|open| {
-            !(open.handle == handle
-                && open.protocol == guid
-                && open.agent_handle == agent_handle
-                && open.controller_handle == controller_handle)
-        });
-        if efi_state.open_protocols.len() == previous_len {
-            Status::NOT_FOUND
-        } else {
-            Status::SUCCESS
-        }
-    })
-}
-
-extern "efiapi" fn open_protocol_information(
-    handle: Handle,
-    protocol: *mut Guid,
-    entry_buffer: *mut *mut efi::OpenProtocolInformationEntry,
-    entry_count: *mut usize,
-) -> Status {
-    log::debug!("BS.OpenProtocolInformation(handle={:?})", handle);
-
-    if handle.is_null() || protocol.is_null() || entry_buffer.is_null() || entry_count.is_null() {
-        return Status::INVALID_PARAMETER;
-    }
-
-    let guid = unsafe { *protocol };
-    let record_count = with_tables_mut(|efi_state| {
-        let protocol_exists = efi_state.handles[..efi_state.handle_count]
-            .iter()
-            .any(|entry| {
-                entry.handle == handle
-                    && entry.protocols[..entry.protocol_count]
-                        .iter()
-                        .any(|entry| entry.guid == guid)
-            });
-        protocol_exists.then(|| {
-            efi_state
-                .open_protocols
-                .iter()
-                .filter(|open| open.handle == handle && open.protocol == guid)
-                .count()
-        })
-    });
-    let Some(record_count) = record_count else {
-        return Status::NOT_FOUND;
-    };
-
-    if record_count == 0 {
-        unsafe {
-            *entry_buffer = core::ptr::null_mut();
-            *entry_count = 0;
-        }
-        return Status::SUCCESS;
-    }
-
-    let size = record_count * core::mem::size_of::<efi::OpenProtocolInformationEntry>();
-    let buffer = match allocator::allocate_pool(MemoryType::BootServicesData, size) {
-        Ok(buffer) => buffer as *mut efi::OpenProtocolInformationEntry,
-        Err(status) => return status,
-    };
-    with_tables_mut(|efi_state| {
-        for (index, open) in efi_state
-            .open_protocols
-            .iter()
-            .filter(|open| open.handle == handle && open.protocol == guid)
-            .enumerate()
-        {
-            unsafe {
-                buffer.add(index).write(efi::OpenProtocolInformationEntry {
-                    agent_handle: open.agent_handle,
-                    controller_handle: open.controller_handle,
-                    attributes: open.attributes,
-                    open_count: open.open_count,
-                });
-            }
-        }
-    });
-    unsafe {
-        *entry_buffer = buffer;
-        *entry_count = record_count;
-    }
-    Status::SUCCESS
-}
-
-extern "efiapi" fn protocols_per_handle(
-    handle: Handle,
-    protocol_buffer: *mut *mut *mut Guid,
-    protocol_buffer_count: *mut usize,
-) -> Status {
-    log::debug!("BS.ProtocolsPerHandle(handle={:?})", handle);
-
-    if handle.is_null() || protocol_buffer.is_null() || protocol_buffer_count.is_null() {
-        return Status::INVALID_PARAMETER;
-    }
-
-    let efi_state = tables();
-
-    // Find the handle entry
-    let entry = match efi_state.handles[..efi_state.handle_count]
-        .iter()
-        .find(|e| e.handle == handle)
-    {
-        Some(e) => e,
-        None => {
-            log::debug!("  -> NOT_FOUND");
-            return Status::NOT_FOUND;
-        }
-    };
-
-    let count = entry.protocol_count;
-
-    if count == 0 {
-        // No protocols on this handle -- return empty buffer
-        unsafe {
-            *protocol_buffer = core::ptr::null_mut();
-            *protocol_buffer_count = 0;
-        }
-        log::debug!("  -> SUCCESS (0 protocols)");
-        return Status::SUCCESS;
-    }
-
-    // Allocate a single contiguous buffer: array of Guid pointers followed by
-    // the Guid values themselves. Per UEFI spec, the caller frees only the
-    // returned buffer with a single FreePool call, so all data must live in
-    // one allocation.
-    let ptrs_size = count * core::mem::size_of::<*mut Guid>();
-    let guids_size = count * core::mem::size_of::<Guid>();
-    let total_size = ptrs_size + guids_size;
-    let buf = match allocator::allocate_pool(MemoryType::BootServicesData, total_size) {
-        Ok(ptr) => ptr,
-        Err(_) => return Status::OUT_OF_RESOURCES,
-    };
-
-    // Layout: [*mut Guid; count] [Guid; count]
-    let ptr_array = buf as *mut *mut Guid;
-    let guid_array = unsafe { buf.add(ptrs_size) } as *mut Guid;
-
-    for (i, protocol) in entry.protocols.iter().take(count).enumerate() {
-        unsafe {
-            let guid_ptr = guid_array.add(i);
-            *guid_ptr = protocol.guid;
-            *ptr_array.add(i) = guid_ptr;
-        }
-    }
-
-    unsafe {
-        *protocol_buffer = ptr_array;
-        *protocol_buffer_count = count;
-    }
-
-    log::debug!("  -> SUCCESS ({} protocols)", count);
-    Status::SUCCESS
-}
-
-extern "efiapi" fn locate_handle_buffer(
-    search_type: efi::LocateSearchType,
-    protocol: *mut Guid,
-    search_key: *mut c_void,
-    no_handles: *mut usize,
-    buffer: *mut *mut Handle,
-) -> Status {
-    let guid_display = if protocol.is_null() {
-        None
-    } else {
-        Some(GuidFmt(unsafe { *protocol }))
-    };
-
-    log::debug!(
-        "BS.LocateHandleBuffer(type={}, protocol={})",
-        search_type,
-        guid_display
-            .as_ref()
-            .map(|g| g as &dyn core::fmt::Display)
-            .unwrap_or(&"NULL" as &dyn core::fmt::Display)
-    );
-
-    if no_handles.is_null() || buffer.is_null() {
-        log::debug!("  -> INVALID_PARAMETER");
-        return Status::INVALID_PARAMETER;
-    }
-
-    // First, call locate_handle with null buffer to get required size
-    let mut buffer_size: usize = 0;
-    let status = locate_handle(
-        search_type,
-        protocol,
-        search_key,
-        &mut buffer_size as *mut usize,
-        core::ptr::null_mut(),
-    );
-
-    // If no handles found, buffer_size is 0
-    if status == Status::NOT_FOUND {
-        unsafe {
-            *no_handles = 0;
-            *buffer = core::ptr::null_mut();
-        }
-        log::warn!("  -> NOT_FOUND");
-        return Status::NOT_FOUND;
-    }
-
-    // Should get BUFFER_TOO_SMALL with required size
-    if status != Status::BUFFER_TOO_SMALL {
-        log::debug!("  -> {:?} (unexpected from locate_handle)", status);
-        return status;
-    }
-
-    // Calculate number of handles
-    let handle_count = buffer_size / core::mem::size_of::<Handle>();
-
-    // Allocate buffer for handles
-    let alloc_result = allocator::allocate_pool(MemoryType::BootServicesData, buffer_size);
-    let handle_buffer = match alloc_result {
-        Ok(ptr) => ptr as *mut Handle,
-        Err(e) => {
-            log::warn!("  -> OUT_OF_RESOURCES (pool allocation failed: {:?})", e);
-            return Status::OUT_OF_RESOURCES;
-        }
-    };
-
-    // Call locate_handle again with the allocated buffer
-    let status = locate_handle(
-        search_type,
-        protocol,
-        search_key,
-        &mut buffer_size as *mut usize,
-        handle_buffer,
-    );
-
-    if status != Status::SUCCESS {
-        // Free the allocated buffer on failure
-        let _ = allocator::free_pool(handle_buffer as *mut u8);
-        log::debug!("  -> {:?} (second locate_handle call failed)", status);
-        return status;
-    }
-
-    // Return results to caller
-    unsafe {
-        *no_handles = handle_count;
-        *buffer = handle_buffer;
-    }
-
-    log::debug!("  -> SUCCESS ({} handles)", handle_count);
-    Status::SUCCESS
-}
-
-extern "efiapi" fn locate_protocol(
-    protocol: *mut Guid,
-    registration: *mut c_void,
-    interface: *mut *mut c_void,
-) -> Status {
-    if protocol.is_null() || interface.is_null() {
-        return Status::INVALID_PARAMETER;
-    }
-
-    let guid = unsafe { *protocol };
-    log::trace!("BS.LocateProtocol(protocol={})", GuidFmt(guid));
-
-    if !registration.is_null() {
-        return with_tables_mut(|efi_state| {
-            let (index, _, entry) =
-                match next_registered_protocol(efi_state, registration as usize, Some(guid)) {
-                    Ok(next) => next,
-                    Err(status) => return status,
-                };
-            unsafe { *interface = entry.interface };
-            efi_state.protocol_notifies[index].cursor = entry.generation;
-            Status::SUCCESS
-        });
-    }
-
-    let efi_state = tables();
-
-    // Find first handle with this protocol
-    let found = efi_state.handles[..efi_state.handle_count]
-        .iter()
-        .flat_map(|entry| entry.protocols[..entry.protocol_count].iter())
-        .find(|proto| proto.guid == guid);
-
-    if let Some(proto) = found {
-        unsafe { *interface = proto.interface };
-        log::trace!("  -> SUCCESS (interface={:p})", proto.interface);
-        return Status::SUCCESS;
-    }
-
-    log::trace!("  -> NOT_FOUND");
-    Status::NOT_FOUND
-}
-
-type InstallMultipleFn = extern "efiapi" fn(*mut Handle, *mut c_void, *mut c_void) -> Status;
-type UninstallMultipleFn = extern "efiapi" fn(Handle, *mut c_void, *mut c_void) -> Status;
-
-#[cfg(target_arch = "x86_64")]
-core::arch::global_asm!(
-    r#"
-    .global crabefi_install_multiple_protocol_interfaces_entry
-    .type crabefi_install_multiple_protocol_interfaces_entry,@function
-crabefi_install_multiple_protocol_interfaces_entry:
-    mov r10, rsp
-    sub rsp, 40
-    lea rax, [r10 + 40]
-    mov [rsp + 32], rax
-    call crabefi_install_multiple_protocol_interfaces_x64
-    add rsp, 40
-    ret
-
-    .global crabefi_uninstall_multiple_protocol_interfaces_entry
-    .type crabefi_uninstall_multiple_protocol_interfaces_entry,@function
-crabefi_uninstall_multiple_protocol_interfaces_entry:
-    mov r10, rsp
-    sub rsp, 40
-    lea rax, [r10 + 40]
-    mov [rsp + 32], rax
-    call crabefi_uninstall_multiple_protocol_interfaces_x64
-    add rsp, 40
-    ret
-"#
-);
-
-#[cfg(target_arch = "x86_64")]
-unsafe extern "efiapi" {
-    fn crabefi_install_multiple_protocol_interfaces_entry(
-        handle: *mut Handle,
-        protocol: *mut c_void,
-        interface: *mut c_void,
-    ) -> Status;
-    fn crabefi_uninstall_multiple_protocol_interfaces_entry(
-        handle: Handle,
-        protocol: *mut c_void,
-        interface: *mut c_void,
-    ) -> Status;
-}
-
-#[cfg(target_arch = "x86_64")]
-const INSTALL_MULTIPLE_PROTOCOL_INTERFACES: InstallMultipleFn = unsafe {
-    core::mem::transmute::<
-        unsafe extern "efiapi" fn(*mut Handle, *mut c_void, *mut c_void) -> Status,
-        InstallMultipleFn,
-    >(crabefi_install_multiple_protocol_interfaces_entry)
-};
-
-#[cfg(target_arch = "x86_64")]
-const UNINSTALL_MULTIPLE_PROTOCOL_INTERFACES: UninstallMultipleFn = unsafe {
-    core::mem::transmute::<
-        unsafe extern "efiapi" fn(Handle, *mut c_void, *mut c_void) -> Status,
-        UninstallMultipleFn,
-    >(crabefi_uninstall_multiple_protocol_interfaces_entry)
-};
-
-#[cfg(not(target_arch = "x86_64"))]
-const INSTALL_MULTIPLE_PROTOCOL_INTERFACES: InstallMultipleFn = unsafe {
-    core::mem::transmute::<
-        unsafe extern "C" fn(*mut Handle, *mut c_void, *mut c_void, ...) -> Status,
-        InstallMultipleFn,
-    >(install_multiple_protocol_interfaces_c)
-};
-
-#[cfg(not(target_arch = "x86_64"))]
-const UNINSTALL_MULTIPLE_PROTOCOL_INTERFACES: UninstallMultipleFn = unsafe {
-    core::mem::transmute::<
-        unsafe extern "C" fn(Handle, *mut c_void, *mut c_void, ...) -> Status,
-        UninstallMultipleFn,
-    >(uninstall_multiple_protocol_interfaces_c)
-};
-
-const MAX_PROTOCOL_PAIRS: usize = 16;
-
-fn collect_protocol_pair(
-    pairs: &mut Vec<(Guid, *mut c_void)>,
-    protocol: *mut c_void,
-    interface: *mut c_void,
-) -> Result<bool, Status> {
-    if protocol.is_null() {
-        return Ok(false);
-    }
-    if pairs.len() >= MAX_PROTOCOL_PAIRS {
-        return Err(Status::INVALID_PARAMETER);
-    }
-    if pairs.try_reserve(1).is_err() {
-        return Err(Status::OUT_OF_RESOURCES);
-    }
-    pairs.push((unsafe { *(protocol as *const Guid) }, interface));
-    Ok(true)
-}
-
-fn install_multiple_pairs(handle: *mut Handle, pairs: &[(Guid, *mut c_void)]) -> Status {
-    if handle.is_null() || pairs.is_empty() {
-        return Status::INVALID_PARAMETER;
-    }
-
-    let created_handle = unsafe { (*handle).is_null() };
-    let target_handle = if created_handle {
-        let Some(new_handle) = create_handle() else {
-            return Status::OUT_OF_RESOURCES;
-        };
-        unsafe { *handle = new_handle };
-        new_handle
-    } else {
-        unsafe { *handle }
-    };
-
-    for (index, (guid, interface)) in pairs.iter().enumerate() {
-        // No callbacks may observe a partial transaction or acquire opens that
-        // prevent rollback. Publish notifications only after every pair succeeds.
-        let status = install_protocol_internal(target_handle, guid, *interface, false);
-        if status != Status::SUCCESS {
-            let mut rollback_complete = true;
-            for (installed_guid, installed_interface) in pairs[..index].iter().rev() {
-                let rollback_status =
-                    remove_protocol(target_handle, installed_guid, Some(*installed_interface));
-                if rollback_status != Status::SUCCESS {
-                    rollback_complete = false;
-                    log::error!(
-                        "InstallMultipleProtocolInterfaces rollback failed for {} on {:?}: {:?}",
-                        GuidFmt(*installed_guid),
-                        target_handle,
-                        rollback_status
-                    );
-                }
-            }
-            if created_handle && rollback_complete && reclaim_empty_handle(target_handle) {
-                unsafe { *handle = core::ptr::null_mut() };
-            }
-            return status;
-        }
-    }
-    let events: heapless::Vec<efi::Event, MAX_PROTOCOL_NOTIFIES> = {
-        let efi_state = tables();
-        efi_state
-            .protocol_notifies
-            .iter()
-            .filter(|notify| pairs.iter().any(|(guid, _)| *guid == notify.protocol))
-            .map(|notify| notify.event)
-            .collect()
-    };
-    for event in events {
-        signal_event(event);
-    }
-    Status::SUCCESS
-}
-
-fn reclaim_empty_handle(handle: Handle) -> bool {
-    with_tables_mut(|efi_state| {
-        let Some(index) = efi_state.handles[..efi_state.handle_count]
-            .iter()
-            .position(|entry| entry.handle == handle)
-        else {
-            // Removing the final installed protocol already reclaimed it.
-            return true;
-        };
-        if efi_state.handles[index].protocol_count != 0 {
-            return false;
-        }
-        efi_state
-            .handles
-            .copy_within(index + 1..efi_state.handle_count, index);
-        efi_state.handle_count -= 1;
-        efi_state.handles[efi_state.handle_count] = HandleEntry::empty();
-        true
-    })
-}
-
-fn uninstall_multiple_pairs(handle: Handle, pairs: &[(Guid, *mut c_void)]) -> Status {
-    if handle.is_null() || pairs.is_empty() {
-        return Status::INVALID_PARAMETER;
-    }
-    for (index, pair) in pairs.iter().enumerate() {
-        if pairs[..index].contains(pair) {
-            return Status::INVALID_PARAMETER;
-        }
-    }
-
-    let validation = tables();
-    let Some(handle_entry) = validation.handles[..validation.handle_count]
-        .iter()
-        .find(|entry| entry.handle == handle)
-    else {
-        return Status::INVALID_PARAMETER;
-    };
-    for (guid, interface) in pairs {
-        if !handle_entry.protocols[..handle_entry.protocol_count]
-            .iter()
-            .any(|entry| entry.guid == *guid && entry.interface == *interface)
-        {
-            return Status::NOT_FOUND;
-        }
-        if validation.open_protocols.iter().any(|open| {
-            open.handle == handle
-                && open.protocol == *guid
-                && protocol_open_blocks_removal(open.attributes)
-        }) {
-            return Status::ACCESS_DENIED;
-        }
-    }
-
-    // Mutation must not overlap the validation borrow of the handle database.
-    drop(validation);
-    for (guid, interface) in pairs {
-        let status = remove_protocol(handle, guid, Some(*interface));
-        if status != Status::SUCCESS {
-            return status;
-        }
-    }
-    Status::SUCCESS
-}
-
-#[cfg(target_arch = "x86_64")]
-unsafe fn collect_x64_pairs(
-    first_protocol: *mut c_void,
-    first_interface: *mut c_void,
-    second_protocol: *mut c_void,
-    stack_arguments: *const *mut c_void,
-) -> Result<Vec<(Guid, *mut c_void)>, Status> {
-    let mut pairs = Vec::new();
-    if !collect_protocol_pair(&mut pairs, first_protocol, first_interface)? {
-        return Ok(pairs);
-    }
-
-    let mut protocol = second_protocol;
-    let mut arguments = stack_arguments;
-    while !protocol.is_null() {
-        let interface = unsafe { *arguments };
-        collect_protocol_pair(&mut pairs, protocol, interface)?;
-        protocol = unsafe { *arguments.add(1) };
-        arguments = unsafe { arguments.add(2) };
-    }
-    Ok(pairs)
-}
-
-#[cfg(target_arch = "x86_64")]
-#[unsafe(no_mangle)]
-extern "efiapi" fn crabefi_install_multiple_protocol_interfaces_x64(
-    handle: *mut Handle,
-    first_protocol: *mut c_void,
-    first_interface: *mut c_void,
-    second_protocol: *mut c_void,
-    stack_arguments: *const *mut c_void,
-) -> Status {
-    let pairs = match unsafe {
-        collect_x64_pairs(
-            first_protocol,
-            first_interface,
-            second_protocol,
-            stack_arguments,
-        )
-    } {
-        Ok(pairs) => pairs,
-        Err(status) => return status,
-    };
-    install_multiple_pairs(handle, &pairs)
-}
-
-#[cfg(target_arch = "x86_64")]
-#[unsafe(no_mangle)]
-extern "efiapi" fn crabefi_uninstall_multiple_protocol_interfaces_x64(
-    handle: Handle,
-    first_protocol: *mut c_void,
-    first_interface: *mut c_void,
-    second_protocol: *mut c_void,
-    stack_arguments: *const *mut c_void,
-) -> Status {
-    let pairs = match unsafe {
-        collect_x64_pairs(
-            first_protocol,
-            first_interface,
-            second_protocol,
-            stack_arguments,
-        )
-    } {
-        Ok(pairs) => pairs,
-        Err(status) => return status,
-    };
-    uninstall_multiple_pairs(handle, &pairs)
-}
-
-#[cfg(not(target_arch = "x86_64"))]
-unsafe extern "C" fn install_multiple_protocol_interfaces_c(
-    handle: *mut Handle,
-    first_protocol: *mut c_void,
-    first_interface: *mut c_void,
-    mut arguments: ...
-) -> Status {
-    let mut pairs = Vec::new();
-    match collect_protocol_pair(&mut pairs, first_protocol, first_interface) {
-        Ok(true) => {}
-        Ok(false) => return Status::INVALID_PARAMETER,
-        Err(status) => return status,
-    }
-    loop {
-        let protocol: *mut c_void = unsafe { arguments.next_arg() };
-        if protocol.is_null() {
-            break;
-        }
-        let interface: *mut c_void = unsafe { arguments.next_arg() };
-        if let Err(status) = collect_protocol_pair(&mut pairs, protocol, interface) {
-            return status;
-        }
-    }
-    install_multiple_pairs(handle, &pairs)
-}
-
-#[cfg(not(target_arch = "x86_64"))]
-unsafe extern "C" fn uninstall_multiple_protocol_interfaces_c(
-    handle: Handle,
-    first_protocol: *mut c_void,
-    first_interface: *mut c_void,
-    mut arguments: ...
-) -> Status {
-    let mut pairs = Vec::new();
-    match collect_protocol_pair(&mut pairs, first_protocol, first_interface) {
-        Ok(true) => {}
-        Ok(false) => return Status::INVALID_PARAMETER,
-        Err(status) => return status,
-    }
-    loop {
-        let protocol: *mut c_void = unsafe { arguments.next_arg() };
-        if protocol.is_null() {
-            break;
-        }
-        let interface: *mut c_void = unsafe { arguments.next_arg() };
-        if let Err(status) = collect_protocol_pair(&mut pairs, protocol, interface) {
-            return status;
-        }
-    }
-    uninstall_multiple_pairs(handle, &pairs)
-}
-
-extern "efiapi" fn calculate_crc32(data: *mut c_void, data_size: usize, crc32: *mut u32) -> Status {
-    if data.is_null() || crc32.is_null() || data_size == 0 {
-        return Status::INVALID_PARAMETER;
-    }
-
-    let slice = unsafe { core::slice::from_raw_parts(data as *const u8, data_size) };
-    let result = crc32::calculate(slice);
-    unsafe { *crc32 = result };
-    Status::SUCCESS
-}
-
-extern "efiapi" fn copy_mem(destination: *mut c_void, source: *mut c_void, length: usize) {
-    if destination.is_null() || source.is_null() {
-        return;
-    }
-
-    unsafe {
-        core::ptr::copy(source as *const u8, destination as *mut u8, length);
-    }
-}
-
-extern "efiapi" fn set_mem(buffer: *mut c_void, size: usize, value: u8) {
-    if buffer.is_null() {
-        return;
-    }
-
-    unsafe { core::slice::from_raw_parts_mut(buffer as *mut u8, size).fill(value) };
-}
-
-// ============================================================================
-// Helper Functions
-// ============================================================================
-
-use super::guid_fmt::GuidFmt;
-
-// GUID name lookup table has been extracted to efi/guid_fmt.rs
-// (was ~350 lines of GUID-to-name mappings)
-/// Create a new handle and register it
-pub fn create_handle() -> Option<Handle> {
-    with_tables_mut(|efi_state| {
-        if efi_state.handle_count == efi_state.handles.len()
-            && efi_state.handles.try_reserve(1).is_err()
-        {
-            return None;
-        }
-
-        let handle = efi_state.next_handle as *mut c_void;
-        efi_state.next_handle += 1;
-
-        if efi_state.handle_count == efi_state.handles.len() {
-            efi_state.handles.push(HandleEntry::empty());
-        }
-        let idx = efi_state.handle_count;
-        efi_state.handles[idx].handle = handle;
-        efi_state.handles[idx].protocol_count = 0;
-        efi_state.handle_count += 1;
-
-        Some(handle)
-    })
-}
-
-/// Driver, child-controller and exclusive relationships require an explicit close.
-fn protocol_open_blocks_removal(attributes: u32) -> bool {
-    attributes
-        & (efi::OPEN_PROTOCOL_BY_DRIVER
-            | efi::OPEN_PROTOCOL_BY_CHILD_CONTROLLER
-            | efi::OPEN_PROTOCOL_EXCLUSIVE)
-        != 0
-}
-
-/// Remove one protocol interface and reclaim an empty handle slot.
-fn remove_protocol(handle: Handle, guid: &Guid, interface: Option<*mut c_void>) -> Status {
-    with_tables_mut(|efi_state| {
-        if efi_state.open_protocols.iter().any(|open| {
-            open.handle == handle
-                && open.protocol == *guid
-                && protocol_open_blocks_removal(open.attributes)
-        }) {
-            return Status::ACCESS_DENIED;
-        }
-
-        let Some(handle_index) = efi_state.handles[..efi_state.handle_count]
-            .iter()
-            .position(|entry| entry.handle == handle)
-        else {
-            return Status::NOT_FOUND;
-        };
-        let entry = &mut efi_state.handles[handle_index];
-        let Some(protocol_index) =
-            entry.protocols[..entry.protocol_count]
-                .iter()
-                .position(|entry| {
-                    entry.guid == *guid && interface.is_none_or(|iface| entry.interface == iface)
-                })
-        else {
-            return Status::NOT_FOUND;
-        };
-
-        efi_state
-            .open_protocols
-            .retain(|open| open.handle != handle || open.protocol != *guid);
-        entry
-            .protocols
-            .copy_within(protocol_index + 1..entry.protocol_count, protocol_index);
-        entry.protocol_count -= 1;
-        entry.protocols[entry.protocol_count] = ProtocolEntry::empty();
-
-        if entry.protocol_count == 0 {
-            efi_state
-                .handles
-                .copy_within(handle_index + 1..efi_state.handle_count, handle_index);
-            efi_state.handle_count -= 1;
-            efi_state.handles[efi_state.handle_count] = HandleEntry::empty();
-        }
-
-        Status::SUCCESS
-    })
-}
-
-/// Install a protocol on an existing handle
-pub fn install_protocol(handle: Handle, guid: &Guid, interface: *mut c_void) -> Status {
-    install_protocol_internal(handle, guid, interface, true)
-}
-
-/// Install an instance, optionally deferring notification until a transaction commits.
-fn install_protocol_internal(
-    handle: Handle,
-    guid: &Guid,
-    interface: *mut c_void,
-    notify: bool,
-) -> Status {
-    let (status, notify_events) = with_tables_mut(|efi_state| {
-        if let Some(entry) = efi_state.handles[..efi_state.handle_count]
-            .iter_mut()
-            .find(|e| e.handle == handle)
-        {
-            // Check if protocol already installed
-            if entry.protocols[..entry.protocol_count]
-                .iter()
-                .any(|p| p.guid == *guid)
-            {
-                return (Status::INVALID_PARAMETER, heapless::Vec::new());
-            }
-
-            if entry.protocol_count >= MAX_PROTOCOLS_PER_HANDLE {
-                return (Status::OUT_OF_RESOURCES, heapless::Vec::new());
-            }
-
-            let Some(generation) = efi_state.protocol_generation.checked_add(1) else {
-                return (Status::OUT_OF_RESOURCES, heapless::Vec::new());
-            };
-            efi_state.protocol_generation = generation;
-            entry.protocols[entry.protocol_count] = ProtocolEntry {
-                guid: *guid,
-                interface,
-                generation,
-            };
-            entry.protocol_count += 1;
-            let events = if notify {
-                protocol_notification_events(efi_state, guid)
-            } else {
-                heapless::Vec::new()
-            };
-            return (Status::SUCCESS, events);
-        }
-
-        (Status::INVALID_PARAMETER, heapless::Vec::new())
-    });
-
-    // Signaled outside the state borrow: notify callbacks call back into Boot
-    // Services, and a nested mutable borrow trips the state re-entrancy guard.
-    for event in notify_events {
-        signal_event(event);
-    }
-
-    status
-}
-
-/// Look up a protocol interface on a handle (internal helper).
-///
-/// Returns the interface pointer, or null if not found.
-pub fn get_protocol_on_handle(handle: Handle, guid: &Guid) -> *mut c_void {
-    let efi_state = tables();
-
-    efi_state.handles[..efi_state.handle_count]
-        .iter()
-        .find(|e| e.handle == handle)
-        .and_then(|e| {
-            e.protocols[..e.protocol_count]
-                .iter()
-                .find(|p| p.guid == *guid)
-        })
-        .map_or(core::ptr::null_mut(), |p| p.interface)
-}
-
-#[derive(Clone, Copy)]
-struct DeferredImageMeasurement {
-    pcr_index: u32,
-    event_type: u32,
-    digest_count: usize,
-    digests: [super::tcg::types::TaggedDigest; 5],
-    event_data: *mut u8,
-    event_data_size: usize,
-}
-
-pub(crate) fn serialize_tcg_image_load_event(
-    loaded_image: &pe::LoadedImage,
-    image_link_time_address: u64,
-    device_path_ptr: *const DevicePathProtocol,
-) -> Vec<u8> {
-    let device_path_size = if device_path_ptr.is_null() {
-        0
-    } else {
-        unsafe { super::protocols::device_path::device_path_size(device_path_ptr) }
-    };
-
-    let mut event = Vec::with_capacity(32 + device_path_size);
-    event.extend_from_slice(&loaded_image.image_base.to_le_bytes());
-    event.extend_from_slice(&loaded_image.image_size.to_le_bytes());
-    event.extend_from_slice(&image_link_time_address.to_le_bytes());
-    event.extend_from_slice(&(device_path_size as u64).to_le_bytes());
-    if device_path_size != 0 {
-        let device_path =
-            unsafe { core::slice::from_raw_parts(device_path_ptr as *const u8, device_path_size) };
-        event.extend_from_slice(device_path);
-    }
-    event
-}
-
-/// Measure or defer a PE/COFF image for TCG measured boot.
-///
-/// Driver images are measured immediately. Application image digests and event
-/// data are precomputed here so `StartImage()` can log them after ReadyToBoot.
-fn measure_pe_image_for_tcg(
-    pe_data: &[u8],
-    loaded_image: &pe::LoadedImage,
-    device_path: *const DevicePathProtocol,
-) -> Option<DeferredImageMeasurement> {
-    use super::tcg::measured_boot::{measure_pe_image_all, precompute_pe_image_digests_all};
-    use super::tcg::types::*;
-
-    let headers = pe::parse_headers(pe_data).ok()?;
-    let subsystem = headers.subsystem();
-    let (pcr_index, event_type) = match subsystem {
-        10 => (4, EV_EFI_BOOT_SERVICES_APPLICATION),
-        11 => (2, EV_EFI_BOOT_SERVICES_DRIVER),
-        12 => (2, EV_EFI_RUNTIME_SERVICES_DRIVER),
-        _ => (4, EV_EFI_BOOT_SERVICES_APPLICATION),
-    };
-
-    let event_data =
-        serialize_tcg_image_load_event(loaded_image, headers.preferred_image_base(), device_path);
-
-    if subsystem != 10 {
-        if let Err(e) = measure_pe_image_all(pcr_index, event_type, pe_data, &event_data) {
-            log::warn!("Failed to measure PE image: {:?}", e);
-        }
-        return None;
-    }
-
-    let (digest_count, digests) = (match precompute_pe_image_digests_all(pe_data) {
-        Ok(result) => result,
-        Err(e) => {
-            log::warn!("Failed to precompute PE image measurement: {:?}", e);
-            None
-        }
-    })?;
-
-    let event_data_size = event_data.len();
-    let event_data_ptr =
-        match allocator::allocate_pool(MemoryType::BootServicesData, event_data_size) {
-            Ok(ptr) => ptr,
-            Err(status) => {
-                log::warn!(
-                    "Failed to allocate deferred PE measurement event data: {:?}",
-                    status
-                );
-                return None;
-            }
-        };
-    unsafe {
-        // SAFETY: `event_data_ptr` points to `event_data_size` bytes just
-        // allocated above, and `event_data` has exactly that many initialized bytes.
-        core::ptr::copy_nonoverlapping(event_data.as_ptr(), event_data_ptr, event_data_size);
-    }
-
-    Some(DeferredImageMeasurement {
-        pcr_index,
-        event_type,
-        digest_count,
-        digests,
-        event_data: event_data_ptr,
-        event_data_size,
-    })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn create_event_rejects_all_runtime_event_types() {
-        assert!(boot_event_type_supported(0));
-        assert!(boot_event_type_supported(EVT_NOTIFY_SIGNAL));
-        assert!(!boot_event_type_supported(EVT_RUNTIME));
-        assert!(!boot_event_type_supported(
-            EVT_SIGNAL_VIRTUAL_ADDRESS_CHANGE
-        ));
-    }
-
-    /// Regression test for SignalEvent on static events: exercises the same
-    /// `signal_event_entry` helper `signal_event` calls, so switching the
-    /// implementation back to `dynamic_event_id_for_handle` fails this test
-    /// instead of silently regressing to INVALID_PARAMETER.
-    #[test]
-    fn signal_helper_accepts_static_events() {
-        let mut events = [EventEntry::empty(); MAX_EVENTS];
-        let (event_id, notify) = signal_event_entry(&mut events, KEYBOARD_EVENT_ID as efi::Event)
-            .expect("SignalEvent must accept the static keyboard event");
-        assert_eq!(event_id, KEYBOARD_EVENT_ID);
-        assert!(notify.is_none());
-        assert!(events[KEYBOARD_EVENT_ID].signaled);
-        #[cfg(feature = "ui")]
-        {
-            let (event_id, _) = signal_event_entry(&mut events, POINTER_EVENT_ID as efi::Event)
-                .expect("SignalEvent must accept the static pointer event");
-            assert_eq!(event_id, POINTER_EVENT_ID);
-            assert!(events[POINTER_EVENT_ID].signaled);
-        }
-    }
-
-    #[test]
-    fn signal_helper_rejects_invalid_handles() {
-        let mut events = [EventEntry::empty(); MAX_EVENTS];
-        assert_eq!(
-            signal_event_entry(&mut events, core::ptr::null_mut()).unwrap_err(),
-            Status::INVALID_PARAMETER
-        );
-    }
-
-    #[test]
-    fn signal_helper_returns_notify_signal_callbacks() {
-        extern "efiapi" fn notify(_event: efi::Event, _context: *mut c_void) {}
-        let mut events = [EventEntry::empty(); MAX_EVENTS];
-        let event_id = find_free_event_slot(&events).unwrap();
-        let generation = next_event_generation(events[event_id].generation);
-        events[event_id] = EventEntry {
-            in_use: true,
-            generation,
-            event_type: EVT_NOTIFY_SIGNAL,
-            notify_tpl: 0,
-            signaled: false,
-            is_keyboard_event: false,
-            notify_function: Some(notify),
-            notify_context: core::ptr::null_mut(),
-            event_group: None,
-            timer_type: TimerType::Cancel,
-            timer_trigger_time: 0,
-            timer_deadline_tsc: 0,
-        };
-        let handle = event_handle(event_id, generation);
-        let (signaled_id, callback) =
-            signal_event_entry(&mut events, handle).expect("dynamic notify event must signal");
-        assert_eq!(signaled_id, event_id);
-        assert!(callback.is_some());
-        assert!(events[event_id].signaled);
-    }
-
-    #[test]
-    fn static_events_are_not_dynamic_handles() {
-        let events = [EventEntry::empty(); MAX_EVENTS];
-        assert_eq!(
-            dynamic_event_id_for_handle(&events, KEYBOARD_EVENT_ID as efi::Event),
-            None
-        );
-        #[cfg(feature = "ui")]
-        assert_eq!(
-            dynamic_event_id_for_handle(&events, POINTER_EVENT_ID as efi::Event),
-            None
-        );
-    }
-
-    #[test]
-    fn close_rejects_closed_events_and_double_close() {
-        let mut events = [EventEntry::empty(); MAX_EVENTS];
-        let event_id = find_free_event_slot(&events).unwrap();
-        let generation = next_event_generation(events[event_id].generation);
-        events[event_id].in_use = true;
-        events[event_id].generation = generation;
-        let handle = event_handle(event_id, generation);
-
-        assert!(close_dynamic_event(&mut events, handle));
-        assert_eq!(event_id_for_handle(&events, handle), None);
-        assert!(!close_dynamic_event(&mut events, handle));
-    }
-
-    #[test]
-    fn closed_event_slots_are_reused_without_reusing_handles() {
-        let mut events = [EventEntry::empty(); MAX_EVENTS];
-        let first = find_free_event_slot(&events).unwrap();
-        let first_generation = next_event_generation(events[first].generation);
-        events[first].in_use = true;
-        events[first].generation = first_generation;
-        let stale_handle = event_handle(first, first_generation);
-        assert_eq!(event_id_for_handle(&events, stale_handle), Some(first));
-        assert_eq!(find_free_event_slot(&events), Some(first + 1));
-
-        assert!(close_dynamic_event(&mut events, stale_handle));
-        let second_generation = next_event_generation(events[first].generation);
-        events[first].in_use = true;
-        events[first].generation = second_generation;
-        let replacement_handle = event_handle(first, second_generation);
-
-        assert_ne!(stale_handle, replacement_handle);
-        assert_eq!(event_id_for_handle(&events, stale_handle), None);
-        assert_eq!(
-            event_id_for_handle(&events, replacement_handle),
-            Some(first)
-        );
-    }
-}
